Move perseus styles to perseus repo.
Summary:
Move relevant styles out of webapp and into the perseus repo. This
will make it nicer for development of perseus.

Additionally, this makes a shell script for copying styles back and forth.

Test Plan:
View perseus test.html and the local webapp to ensure styles
are consistent. (testrender.html is not caught up yet.)

Reviewers: alpert, alex

CC: eater

Differential Revision: http://phabricator.khanacademy.org/D2636