Add basic calculator support to perseus
Summary:
This adds support for perseus opening and closing the calculator on a
per-item basis, and adds a checkbox for enabling and disabling the calculator
in the editor.

Test Plan:
Make a new item. Check the "show calculator" checkbox. Add the item
to an exercise with few other (non-calculator) items in it. Go through the
exercise. Ensure that the calculator shows up when that item is shown. Ensure
that the calculator goes away when the next item is shown. Ensure that the
calculator is then cleared the next time the calculator-enabled item shows up.

Reviewers: alpert, eater

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2681