Fix styling of orderer's stack elements
Summary:
To fix the stack elements, we need to add another element that creates a
z-index context around the entire card. We do this by adding a new
element around the card, and splitting up the styles between the two
evenly.

Test Plan:
Open a perseus exercise with an orderer in a real exercise view (not
the editor or locally). Make sure that the stacks display correctly, that mouse
offsets work correctly when dragging, and that The widths of elements stay
appropriate.

Reviewers: alpert

Differential Revision: http://phabricator.khanacademy.org/D2894