Fix focusing behavior in orderer
The extra key on the first inputs was causing React to think that it should recreate the input and thus lose the focus. The focusExtra method was also leftover and wasn't necessary at all, so I got rid of it.

Test Plan:
Add a card by typing into the last box in the orderer, make sure the box doesn't lose focus after the first character. Backspace back over that text and again don't lose focus.

Auditors: jack