Perseus editor hint alignment
Summary:
In the perseus assessment item editor, align the hint previews with the
hint edit box, so that changes to hints are immediately visible.
Implemented with changes to react components and css tables.

Test Plan:
Edit an item with the item editor; ensure that no javascript
errors occur, that you can edit, reorder, and save hints, and that hint
previews are visually aligned with their editors.

Reviewers: alex

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4322