Add protractor widget
Test Plan: Create an item with protractor in test.html, modify the image and image position. Use JSON output to test in testrender.html. Verify that protractor doesn't interfere with scoring by creating an item with a protractor and an input-number and scoring under various conditions: input-number is blank is scored as empty; input-number has wrong answer is scored as not-empty-but-wrong; input-number has right answer is scored as not-empty-but-right.

Reviewers: alex

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4432