Not-completely broken mobile support for widgets that were previously completely broken
Summary:
This is horribly and embarrasingly awful. But slightly less so than things not working at all. I spent most of the day coming to the conclusion that this is really hard to solve w/ react in the way. Basically someone needs to reimplement jquery.mobile.vmouse's functionality on top of React's SyntheticEvent system. Maybe next FIXIT... Until then, I think this is a slight improvement.

Now you have to tap the thing you want to drag, then drag it. (Or frustratingly try to drag it twice.) But I suppose that's better than not being able to use these exercises at all on mobile.

Test Plan: Hit up testrender.html on an iPad with orderer and categorizer. Cry.

Reviewers: alex, alpert

Reviewed By: alpert

CC: alpert, laura

Differential Revision: http://phabricator.khanacademy.org/D4873