Fix transformer reset button being hidden by image
Summary:
The reset button could be covered or partially covered by a background
image on the graph. This commit sets the reset button's z-index to 1 to
prevent this.

I tried playing around with setting the image's z-index to -1 instead,
but this just resulted in the image not showing.

Test Plan:
add a large background image that covers the reset button.
verify that the reset button is still clickable

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D5031