Buttons to activate transformer tools
Summary:
Adds a toolbar of buttons to the transformer widget to activate various
transforms, instead of showing all the transforms on the graph at once.

Test Plan:
Visit test.html and create a question. make sure all the tools are
useable and it is gradable
Visit a previously saved question and check that it is completable

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D5088