Make transformer's shape editor snap to grid
Summary:
This commit makes the points and polygons in the transformer
widget's starting location shape editor snap to the background
grid, as requested by Tomer.

It's unclear if we will need an option to allow polygons to not
snap to the grid. Adding this would not be difficult, but maybe
we should see how things go with snap before allowing non-snap
(again).

This commit should not modify the coordinates of previously
created shapes that were not on the grid, unless the shape
is modified.

Test Plan:
Visit test.html, drag points or the polygon in the shape
editor. Also test with lines or line segments.

Edit test.html by hand to put the points at strange positions,
and make sure this doesn't affect the saved position of the points
in the non-editor window. This is important for previously saved
problems.

Reviewers: alex

Reviewed By: alex

Differential Revision: http://phabricator.khanacademy.org/D5127