Transformer: remove extra graphie resets
Summary:
Now that we have better handling of the this.props.transformations list,
we don't need to call setupGraphie when this.props.transformations
changes, as that is handled by this.setTransformations(). Removing these
extra calls should increase the widget's performance in formal mode,
where there will be frequent edits to the transformation list.

Now that we aren't resetting graphie on transformation list changes, we
also should hide any visible tools when the reset button is pressed, as
those won't be hidden automatically anymore.

Test Plan:
Visit test.html. Step through and make sure shouldSetupGraphie returns
false when a transformation is added or when the reset button is
pressed.

Make sure the reset button also hides any tools currently on-screen.

Reviewers: alex

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D5191