Replace Transformer reset button with undo button
Summary:
Places an undo button in the lower-right hand corner of the transformer,
and removes the reset button from on-top of the transformer.

Test Plan:
visit test.html, add some transformations,
undo. Make sure undo removes the last transformation and un-applies it
whether there is only a single transformation or multiple, and make sure
it hides the current UI tool but otherwise has no affect if there are no
transformations.

Reviewers: alex

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D5265