make: do not continue on jsx compilation error
Summary:
If there is a jsx compilation error, we should stop our makefile and not
allow a subperseus to occur.

Since jsx always seems to return 0, even on error, the Gruntfile now
checks for the non-empty existance of all the srcFiles after running
jsx, and errors if some are missing, halting the Makefile.

Test Plan:
add a jsx compilation error to src/widgets/transformer.jsx
type 'make all'
verify that concatenation and './operations put' do not appear in the
output, and that build/perseus.js does not exist

Reviewers: alex, alpert

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D5452