Transformer: Fix batman sign points
Summary:
graphie.addMovableLineSegment is evil, and this is a quick hack to fight
that evil at the scene of the crime rather than addressing its
deep-seated childhood psychological issues (yet).

Test Plan:
Reflect the polygon, then translate the polygon.
Verify that the points of the polygon are perfect circles, not batman
signs.

Auditors: alex