Change the colors for fixed transformation tools
Summary:
Changes the colors for points defining fixed-position transformation
tools, since the orange implies that they are movable.

Gray was the best color I came up with, but maybe it's confusing with
the solution shape shown in gray?

I might also like to highlight the gray shapes as orange when you hover
over the movable part of them (like the old dilation tool did, before
it was always orange).

Test Plan:
 * Make a transformer in test.html
 * Set all tools to fixed
 * View all the tools

Reviewers: eater, khirasaki

Reviewed By: eater

CC: alpert, alex

Differential Revision: http://phabricator.khanacademy.org/D5696