Remove integer-clamping on graphie scale in InteractiveGraph
Summary:
This fixes background image alignment bugs reported by Sandy and Tomer.
This change affects scaleFromExtent, introduced in D3110

This change is scary in that:

1) It may break alignment in any assessment items that unintentionally
relied on the wacky scale returned by scaleFromExtent. This could
potentially affect any exercise with an interactive graph and a
background image.

2) Many graphie-to-png templates make
use of this function, and changing it will make old graphie-to-png
images not re-render identically if they are recreated from their js
source. However, this change will make the "desired size" of these
images more accurate. Previously, specifying size=400 would get you an
image with size close to 400, now you will get an image that is exactly
400 pixels.

Test Plan:
Change the y scale from -10,100 to -5,50 step 10 in:
https://www.khanacademy.org/devadmin/content/items/xca2cd29b1e2d1fcb?revision=79bc702ee8734f1b59b13aab93257f68c9847631
Check that the honeycomb is aligned in the same place for both edit and
preview mode in:
https://www.khanacademy.org/devadmin/content/items/xcade801669fd6399?revision=8ec19afc90da3f8f941c5d92fda0957891ea0110

Reviewers: alex, alpert

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D4782