Add angleOffsetDeg to interactive-graph's angle
Summary:
Add an offset degree amount to the movableangle subwidget in
interactive-graph, so that question writers can better use
snapDegrees to line up the angle with background images.

Note: Not all angles can be snapped to, as the degree measure of
the angle must still be a multiple of snapDegrees. When I changed this,
it became really awkward to use because of interactions with how the
angle wraps around (when you hold the same point and drag it around
the angle)

Test Plan:
 * Set snapDegrees to 16 and an offset of 0, and make sure you can
   line them up with the background image.
 * Change the offset to 8, and make sure you cannot line them up.

Reviewers: alex, justinjaffray

Reviewed By: alex

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6128