Broke out TextListEditor as a reusable component and improved its interactions.
Summary:
You can now use backspace and enter within the TextListEditor to easily insert
and delete anywhere in the list, instead of just at the end.

Also moved a couple of InfoTips in the Orderer's editor to better locations.

Test Plan: Tested in both the existing Orderer and upcoming Matcher widgets.

Reviewers: jack

Reviewed By: jack

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6034