Add field to change the snapping of a graph
Summary:
This commit adds a field to change the step at which points snap to the
grid. The two options that made most sense to me were to either define
the snapping in terms of the number of grid points, or to define it in
absolute units. After discussion it was decided that absolute units made
the most sense.

When the gridStep is automatically changed, this value is updated to be
half of it (the old value) as a reasonable default.

Also fixes a bug with setting the gridStep based on the range

Test Plan:
* Open test.html and have an interactive graph
* Change the snap step to some other value
* See that the points now snap to the value you specified
* Try changing the range and grid steps
* See the snap step change in tandem with them

Reviewers: alex, jack

Reviewed By: jack

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D6257