Get tests running again
Summary:
The tests for Perseus haven't been running for a while, so this commit
gets them working again.

Some are still failing, I have a follow-up commit that addresses this,
this commit is just for running and specifying the correct behaviour.

I'd like a double check that the tests I changed actually have the
desired behaviour now, just to be sure.

Test Plan:
* Run the server with `make server`
* Open test/index.html
* See all the tests run (but not pass)

Reviewers: alex, jack

Reviewed By: alex

CC: eater, alpert

Differential Revision: http://phabricator.khanacademy.org/D6281