Add international decimal_separator support to expressions
Summary:
This isn't a simple input.replace(/,/./g) since we don't
want english-speaking students who enter `40,000` to have
it treated as 40, but decided it might be awkward to force
students in other countries to use commas, so a `.` is
always supported. Unfortunately, a parsed `,` still renders
as a '.', but the question will at least grade correctly.

TODO: make it render correctly, too

Test Plan:
 * subperseus into webapp
 * make a question with the answer `1.2`
 * preview that question with ?lang=pt
 * enter `1,2`.
 * make sure that the expression renders (it will display `1.2`)
 * make sure that check answer marks the question correct

Reviewers: alex

Reviewed By: alex

CC: justinjaffray, alpert

Differential Revision: http://phabricator.khanacademy.org/D6440