Add first bit of question diffing
Summary:
This commit adds the first bit of a way to view diffs for questions, as
a React component that shows the difference in the sources.

It includes a js text diffing library used to diff the actual content of
the questions. I also have some code to diff objects (for widgets) but I
decided to leave that out in this commit in the interest of keeping it a
bit smaller.

{F12939}

Test Plan: * Check out difftest.html

Reviewers: jack, alex

Reviewed By: jack

CC: alpert, joel

Differential Revision: http://phabricator.khanacademy.org/D6618