Fix Perseus text-editor word wrap
Summary:
When it overflows, it should expand the textarea so there is no scroll bar ever.
From
{F13566}
To
{F13565}
I've never used this css property before, but it //seems// to be in the CSS3 specification, and for the purpose it's used for here, it appears to be supported in all the browsers.

Test Plan:
1) Open up test.html, and in any perseus edtior (question box, hint, etc.) write some really long string, i.e.
"asdfghjklqwertyuiopzxcvbnmwertyuiopasdfghjklzxcvbnmasdfghjklqwertyuiopsdfghjkzxcvbnmwertyuisdfghasdfghjklwertyuiozxcvbnmj"
2) Test in Chrome, FF, Safari, IE, Opera

Reviewers: alpert

Reviewed By: alpert

CC: tabitha, alpert

Differential Revision: http://phabricator.khanacademy.org/D6990