Refactor MovablePoint options
Summary:
Moves default function options out of movable-point.js, and
into movable-point-options.js. This makes both files much shorter,
and makes the options more libraryesque.

This also nicely adds an "standard" property to every
MovablePoint.add/MovablePoint.draw/etc, which is the default
value for that option, so that if you want to only add functionality
to a movablePoint, instead of destroying the default functionality,
you can do that by specifying:
    draw: [MovablePoint.draw.standard, function() {
        // MovablePoint has already had the usual drawing occur
		// when this code happens
    }]
without having to know that the default for MovablePoint.draw is
[MovablePoint.draw.basic, MovablePoint.draw.highlight]

Note: This commit does not address a better way of normalizing the
function options; right now that is ugly.

Test Plan: * Make an interactive-graph with two points, and move them around

Reviewers: justinjaffray

Reviewed By: justinjaffray

CC: alex, alpert

Differential Revision: http://phabricator.khanacademy.org/D7099