Add Clues to Radio (Multiple Choice)
Summary:
Clues are (ideally) common mistakes or misconceptions that users may have. For example, not simplifying or using a "x" as a "*" are two clues common enough that they're currently automated. This extends that functionality to content creators to address more particular-question concerns. Hopefully, this results in less frustration, greater engagement, and better learning.

Clues, in general, dictate a `condition` under which they are shown, a `message` that will be shown to the user, and (atm) a `correct` and `empty` that determines whether the guess is marked correct, ungraded, or wrong (never use true/true). Thus, you could have (3x)/(4y) be the expected answer, but also mark 3x/(4y) correct and 3x/4y as ungraded with a message asking the user to further clarify the denominator. Or if the student distributes 30(1+2x) as 31+32x, there is obviously a message that could really help that student out, even if its ultimately marked wrong.

For radio in particular, this process is streamlined. The condition is whether or not the user picked that option. If it's not the correct answer, it's marked incorrect. There is no ungraded besides actually being empty. Because multiselect is all bundled up in here, if they choose multiple options, they get shown all of those clues at once.

This diff intends to be landed this week, but it's not the _final_ version. But if you waited for the final version, it would be a thousand line diff with unseparated concerns. It would be better if clues marked correct were green rather than blue; it would better if they were in a pop-out box altogether; it would be better if the alignment of clues with the choice wasn't manually set but rather used inline-block. And it would be a lot better if you could see the most common answers to the question somewhere. All in due time.

For now, we have {F13701}.

Test Plan:
1) Look at screenshot, admire.
2) Open test.html, it's already on radio (which will be deleted before landing)
3) Choose an option on the right side (renderer), and press "Score" in top-top-top-left of screen; should show clue in blue.
4) Be annoyed that toggling the clue doesn't actually shut it off, read comments in code for more discussion on this.

Reviewers: alex

Reviewed By: alex

CC: alpert, eater, tabitha

Differential Revision: http://phabricator.khanacademy.org/D6898