Add this.change() and an example-widget
Summary:
Adds Changeling, to make dealing with this.props.onChange
simple, easy, nice, and fun.

Also, I wanted to play around with an idea for widgets, and we didn't
have a really simple base widget to start that from, so I decided
to write one. I think such a widget would also be helpful to point
potential contributors to.

Test Plan:
 * add an example-widget
 * test grading of things equal to the solution, not equal to the
   solution, or emtpy

Reviewers: brianmerlob, justinjaffray, alex

Reviewed By: alex

CC: joel, alpert

Differential Revision: http://phabricator.khanacademy.org/D7302