Fix feedback on image dragging.
This was broken in Chrome and (I think) Firefox. I guess dragenter
doesn't have access to file data. Not sure when that happened. I think
it's still possible to get the file types at that point but in the
interest of shipping I'm committing this now.

Test Plan:
* make all
* drag an image into a textarea in Perseus One
* the textarea gets lighter

Auditors: brianmerlob