move widgets above scratchpad
Summary:
This commit moves the perseus widgets above the scratchpad, and enables
the scratchpad for all perseus items.

the scratchpad is at z-index 1;
this moves widgets to z-index 2 (.above-scratchpad)
this moves interactive parts of widgets to z-index 3 (.interactive-component)
z-index 4 is generally for things currently being dragged

.above-scratchpad is generally declared explicitly rather than used as a
mixin so that we can highlight their borders when the scratchpad is
enabled.

.blank-background is used to prevent drawing under transparent widgets.
drawing under transparent widgets is particularly annoying because if
you lift up your pen and put it down again, you can't continue drawing,
because you're now clicking on the widget. This is very
counterintuitive, so we simply don't let you see anything you draw under
a widget.

Test Plan:
 * subperseus
 * make all the widgets
 * turn on the scratchpad
 * draw on the scratchpad
 * try the interactive widgets to make sure the css
   didn't break them
 * yay

Reviewers: brianmerlob

Reviewed By: brianmerlob

Subscribers: alpert, eater

Differential Revision: http://phabricator.khanacademy.org/D7552