Have matcher and sorter call onChange
Summary:
Necessary for highlighting of widgets, related to {D7174}, we need a way
for the top-level components to be aware that a widget has been
interacted with. This adds a call to this.props.onChange to the matcher
and the sorter that accomplishes this. Ideally we would pass up an
accurate representation of the state of the widgets, but currently this
does not live in props, so there is no way to effectively use it.

Test Plan:
* Add a matcher and a sorter
* Interact with them
* should un-highlight

Reviewers: jack, alex

Reviewed By: alex

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D7554