Give a message that the last hint is bolded.
Summary:
There has been at least one question about this in the past, and I
wouldn't be surprised if there will be more. It's a little surprising
(and I checked the code to see what was happening the first time I
saw) so I think a notice is warranted.

Test Plan:
* Open perseus
* see {F15391}

Reviewers: alan, jack

Reviewed By: jack

Differential Revision: http://phabricator.khanacademy.org/D7659