Warn the user if they are using a non-square image
Summary:
If plotter gets an image that isn't a square, it stretches it to be a
square. This causes situations like this:

{F15558}

So now when the image isn't a square, the user receives this warning:

{F15559}

The long-term fix for this is probably to either handle non-square
images (but I can't think of a particularly nice way to do that) or to
provide a way to crop the image.

One thing to consider is that we might want to fuzz it a bit to allow
for images that are "square enough", but I'm not sure what an
appropriate tolerance might be.

Test Plan:
* Add a plotter, and use a square image as the 'pic'
  * http://i.imgur.com/Y59UMTZ.png is one
  * Should not get any warning
* Change it to a non-square image
  * http://i.imgur.com/ophEqb8.png is one
  * Should get the noted warning

Reviewers: eater, jack

Reviewed By: jack

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D7812