Fix all widgets being inline-block
Summary:
As a consequence of highlighting widgets (and wrapping them in another
div) they took on the display style of that div, which was set to
inline-block. That led to some weird issues with radios and other things
being inline when they shouldn't.

This commit makes every widget specify its displayMode so that the
widgetContainer can style itself appropriately.

Test Plan:
* Add a radio and some text before it
* Should not be inline-block

Reviewers: jack

Reviewed By: jack

Subscribers: alpert

Differential Revision: http://phabricator.khanacademy.org/D8071