Allow for inline multiple choice answers
Summary:
Adds a checkbox for inline mode:

{F16605}

I'd like to do Kitt's full grid thing where the items themselves are the
checkboxes, but realistically I'm not sure I'll have time to before I
leave. I feel like this gets us 80% of the way there for less work,
although if someone else plans to do the grid way I'm happy to not land
this.

However, they're probably compatible, so we could use this for the time
being and switch to something fancier in the future.

Test Plan:
* https://gist.github.com/justinj/95ab8f1c3a3b01ba0ee4
* Try turning "One Per Line" on and off.

Reviewers: jack

Reviewed By: jack

Subscribers: brianmerlob, khirasaki

Differential Revision: http://phabricator.khanacademy.org/D8486