Widget clearing prop whitelist instead of blacklist
Summary:
We've got some nasty issue on google code where widgets are being
reset when someone clicks the check answer button. I can't repro
this, but this widget clearing code is a bit suspect because if
executed, it could result in those symptoms, and this code has
caused us problems before. It should be safer to use a whitelist
here than a blacklist, because adding props to the renderer
won't later cause widgets to reset. Of course, the opposite
problem could occur, where we add a prop that we would like to
reset widgets when changed (anything related to the item), but
this case seems less likely.

Test Plan:
Test doing an exercise on staging.
Cross fingers.

Auditors: alex