Fix categorizer so removing an item doesn't leave it ungradable
Test Plan:
In local perseus, add a categorizer with 3 categories and 3 items. Inspect the JSON and see this:

```
    "categorizer 1": {
        "type": "categorizer",
        "graded": true,
        "options": {
            "items": [
                "1",
                "2",
                "3"
            ],
            "categories": [
                "A",
                "B",
                "C"
            ],
            "values": [
                0,
                1,
                2
            ]
        }
    }
```

Remove one of the items and reinspect the JSON and verify there are the same number of values as items (previously values would be unchanged from above):

    "categorizer 1": {
        "type": "categorizer",
        "graded": true,
        "options": {
            "items": [
                "1",
                "2"
            ],
            "categories": [
                "A",
                "B",
                "C"
            ],
            "values": [
                0,
                1
            ]
        }
    }

Select the correct answer in the preview area, hit the score button, and verify in the console that the answer is graded as correct (previously it would be scored as incorrect).

Reviewers: joel, jack

Reviewed By: jack

Subscribers: jack

Maniphest Tasks: T2690

Differential Revision: http://phabricator.khanacademy.org/D9961