update movableline to be more like movablepoint in semantics
Summary:
I made a bunch of changes to MovablePoint when landing <Graphie> to make
MovablePoint's semantics make sense with <Graphie>. This diff should
propagate most of those changes to MovableLine.

Test Plan:
Make an interactive-graph and play with the line.
Make sure the visuals all look good and that grading works

Reviewers: eater, charlie

Reviewed By: charlie

Differential Revision: http://phabricator.khanacademy.org/D9983