package.json: Update deps
We were hitting some large subperseus differences because our
react-components dependency was anything >=0.0.3. I updated
us to be using 0.0.8, so that we shouldn't see huge
react components changes when subperseusing depending on
how recently that dev has locally updated react-components.

I dropped the `>=` or `~` from the version so that a subperseus
is more deterministic. We can just up the version requirement
when we need to update things, and then everyone will see
the update.

Test Plan:
 * npm install
 * make test && make jest
 * make server
 * open up perseus and play around

Auditors: joel