Added a sine/cosine option to interactive-graph
Summary:
Eater mentioned today that most of our trig questions only ask students to plot a few
points, but don't enforce much interaction with the actual shape of the graph. The exercises in ke
also seemed somewhat limited, so I added functionality to interactive graph similar to the 'quadratic
function' option which will let students graph trig functions by moving key points around the plane.
The methodology is that you provide the user with two key points: the first point (always left-most,
to preserve the 1-to-1 mapping of point positions to sine functions) indicates the location of the
'origin' of the sine curve, and the second point indicates the location of the closest peak/trough
(these definitions should be immediately obvious after playing around with the widget).
Addresses T2800.

Appropriate questions might be of the form: "Plot 4 * cos(3x - 2) + 5 on the graph below".

Test Plan:
- Run Perseus.
- Create an interactive graph.
- Flip to the 'Trigonometric function' setting.
- Play around with the two points and check that:
    - The curve adjusts to the two points as expected (move them both in every direction).
    - The right-most point can be either a peak or a trough (i.e., can go below the left-most point).
    - The equation string (in the editor pane) is accurate.
    - The 'origin' point is always the left-most.
- You'll also want to check that:
    1. Grading works correctly.
    2. Permalinking works correctly.

Reviewers: eater, jack

Reviewed By: jack

Differential Revision: http://phabricator.khanacademy.org/D10483