Remove `this.props` references from `getDefaultProps`
Summary:
React 0.11 isn't going to be very happy about our deriving
the return value of `getDefaultProps` from `this.props`, so
we're removing references to `this.props`.

We previously used this when we added a new prop such as gridStep
or tickStep, which was previously derived from the old props.
In this situation, we needed old items that were saved
without this prop to be upgraded to use the correct default
based on their range and step, rather than a constant default.
In the future, we will handle this case with prop version
upgrades, instead.

We need to do this in the editor as well so that if a
content creator loads and saves an old question, they
don't change its grid/snap semantics.

Note: we don't really do "the right thing" here for the
transformer, because apparently we never have, so things
must be working reasonably without it (and changing it
at this point would probably break more things than not).

Test Plan:
 * Made an interactive-graph and transformer in test.html
 * Added an interactive-graph with no gridStep or tickStep
   to testrender.html, loaded testrender.html, and made sure
   that it rendered and the defaults for grid and tick step
   made sense based on the range and step.
 * Added the same interactive-graph to test.html, and loaded it,
   making sure the default grid and snap step were correct.

Reviewers: charlie, joel

Reviewed By: charlie, joel

Differential Revision: http://phabricator.khanacademy.org/D11127