Fix NumberInput's check for a valid answer
Summary:
NumberInput tries to check for whether your answer is possible.
If not, it will highlight the "correct x" answer box in red.

However, its logic for doing so doesn't work well with
floating point arithmetic, where mod (%) might return
something like `5.551115123125783e-17`, or `0.999999999999999`.
knumber.isInteger is made to handle these cases, so
we just check if the answer divided by the tick is an integer.

Test Plan:
Make a number-line with range [0, 7/6] and 7 ticks
Set the answer to `1`
See that the answer box is not highlighted in red
Try setting the answer to `1.1`
See that the answer box is highlighted in red

Auditors: michelle