Update number-line widget to allow clicking on the number line to move the point
Summary:
This adds better interaction support for the number line widget. Now when you click anywhere on the widget, it will move the point to the nearest valid spot on the number line, and will also "grab" it so you can fluidly drag it from where you click. Fixes T2914.

To make this happen we added `ref="..."` support to `<Graphie>`'s movable children.

Test Plan:
Play with the number-line widget in perseus, expecting the following behavior:
- You can click+drag the dot to move it to snapped points
- You can click+drag from anywhere on the numberline to move the dot to snapped points
- You can single-click anywhere on the numberline to move the dot to snapped points
- You can't make the dot move to somewhere outside of the line, or to a non-snap-point
- Answers are still accepted/rejected as you expect (i.e. the dot is in the right place)

Reviewers: jack

Reviewed By: jack

Subscribers: charlie, jack

Maniphest Tasks: T2914

Differential Revision: http://phabricator.khanacademy.org/D11477