Graphie: Fix default range and scale overriding
Summary:
With the default values included second they overwrite the values from the widget and break number-line, interactive graph, etc. By moving `this.props.options` to be after the defaults, the values from the widget will take precedence.

Test Plan:
Load test.html
Move the points on the interactive-graph and watch the line move with the points
Add a number-line widget and see that it appears

Auditors: eater jack