Allow markdown in passage-refs
Summary:
Adds an option to make a renderer inline, which we then use
to allow markdown inside of passage-ref-targets.

Long term, we should add a different grammar for a markdown-lite
that we support for passages, and do something differently so that
we can have passage refs that span across paragraphs, but this works
nicely within paragraphs for the short-term, and will allow
question writers to write questions.

Test Plan: tried adding markdown inside {{a passage ref target}}

Reviewers: michelle, charlie

Reviewed By: charlie

Differential Revision: http://phabricator.khanacademy.org/D12113