Add title and footnotes to the passage widget
Summary:
The title field allows question writers to specify an (optional) title for each passage, which is used when there are 2 or more passages in the question (some SAT passage questions ask the reader to compare and contrast multiple passages, and they are usually titled "Passage 1", "Passage 2", ...). The title is added above the passage in bold with the same font and indentation as used in the passage itself.

The footnote section allows the question writer to add superscripts in the passage text (using KaTeX, e.g. `${}^1$` or `${}^*$`) and then add the corresponding footnote below. The footnotes are styled with the same font and indentation as used in the passage, with a smaller font-size and no line numbers.

Test Plan: Add an SAT passage widget and play with the footnotes and title

Reviewers: aria

Reviewed By: aria

Subscribers: emily, alpert, marcos, eater

Differential Revision: http://phabricator.khanacademy.org/D11874