renderer: better handling of invalid widget types
Summary:
Item x66d49f1bab5f8fa4 was crashing on the site because
of references to an "image-widget". This change makes
the renderer code-path resilient to nonexistant widgets
(in addition to the editor, which already worked).

Test plan:
Open localhost:9000/testrender.html, which has an
invalid "image-widget" widget (the type is just
"image" these days).

(the changes in testrender.html are just for testing)

Auditors: michelle