simple-markdown: don't allow defs inside paragraphs
Summary:
previously, using a [def]: like this inside a paragraph
would turn that part of the text into a def (thereby hiding it).
This should just be inline text in the paragraph since it's not
at the block level.

This change is a very, very hacky fix to that. The right
fix involves either parsing a group of defs at once, or
adding block vs inline parameters to the regexes in the
grammar, which we might have to do to comply with
"standard markdown" anyways.

Test plan:
`make test`

Auditors: emily