New Matrix Widget
Summary: The matrix widget allows you to create an interactive matrix "board", where the student can create a matrix (or vector) up to the max size of the board. The user is warned when they enter an incomplete matrix that is the correct size (i.e. with blank cells). You can also specify an optional prefix or suffix for the matrix (e.g. M = [Matrix]). Fixes T3034.

Test Plan:
Create a matrix widget in perseus and try to break it!
Things to watch for:
- Does the matrix grade correct, incorrect, and incomplete as you would expect?
- Do the matrix braces always line up with your cursor and the matrix content?
- Does changing the max matrix size affect the inner matrix answer the way you would expect?

Reviewers: aria

Reviewed By: aria

Subscribers: alex, eater, cameron

Maniphest Tasks: T3034

Differential Revision: http://phabricator.khanacademy.org/D12726