simple-markdown: parse urls inside text as a url
Summary:
The first 'http' string inside text was being parsed as text
rather than the start of a url, causing urls to not be
recognized.

This change breaks the `text` rule on http so that this no
longer happens, and adds a test case (and fixes an old test
case).

 #fixingfuturebugs

Test Plan:
`make test`

Auditors: emily