Add shortcut property to widgets
I made possible to explicitly assign a shortcut to a widget, by adding the 'shortcut' property in the widget's exported object (example included).

The property is not mandatory, so if you don't bother about giving a specific shortcut to a certain widget, one will still be assigned by extracting it from the widget's 'name' property. The 'shortcut' property will have the priority over the automatically generated ones anyway, so we avoid having them accidentally overwritten.