Fix graphie movable factory for forward-compat
When I was testing the changes around descriptors, I didn't test this properly with React 0.11, and it of course fails because the factory was producing an instantiated graphie object instead of the descriptor that the rest of the code now expects. I've replaced the factory with a function that acts similarly to ReactDescriptor.createDescriptor does in 0.12, which is the function that ends up being called by the JSX transform. (In React 0.12, this function is never called so this will still work.)

Test Plan: Made an interaction with a movable point and dragged it around.

Auditors: aria