Graphie: Fix refs
Summary:
We were losing refs added by the new react descriptors.

This change refactors a bit of how we do keys and refs in
<Graphie>; pulling them directly off the descriptor rather
than using the resulting movable object.

This was causing our numberLinePoint ref in number-line
to not be set, which broke clicking on the line to move
the point (since this grabs the ref to call a method on
it).

Test Plan:
 * Make a number-line
 * click somewhere on the line without dragging
 * see no js errors :D

Auditors: alpert