simple-markdown: prevent lists from breaking paragraphs
Summary:
Previously, '*'s or '-'s inside paragraphs would break
them and start a list, which is not what we want.

This change fixes that by taking away the rule where
list items can end with `$` (end of string), and instead
makes all list items end with a '\n'. There is various
rearranging to make this work, but ultimately i think it
makes the list code a bit simpler (all the newlines are
now in the same place, consistently), and since a list
must end with at least one newline, it won't be recognized
when we're parsing inline elements.

Test Plan: `make test`

Reviewers: emily

Reviewed By: emily

Subscribers: michelle

Differential Revision: http://phabricator.khanacademy.org/D13037