widget-prop-blacklist: blacklist questionCompleted
Summary:
It looks like we added the questionCompleted prop to each
widget so that we can do things like the multiple choice
clues correctly (only after the problem has been graded
as something other than "invalid"), but we didn't add
it to the prop blacklist. Because of this, it was being
added to getUserInput()s (previously toJSON) of widgets.
Since this isn't really part of the user input for a
widget, we shouldn't add it there. This commit adds it to
the blacklist so that it stops showing up in problem logs

Test plan:
\#yolo

Auditors: joel