Fix `OnChange` vs `change` bug in grapher widget
Summary: This fixes the JS errors/missing data that occurs when you blur the backgroundImage field or click the ruler/protactor checkboxes in the grapher widget by switching to using `change` instead of `onChange`. This ensures that the widgets props are merged with the changed props instead of being overwritten.

Test Plan:
- Click into/blur the url field (background images aren't supported yet, though!) and notice no JS errors appear
- Check/uncheck the ruler/protactor checkboxes and notice no JS errors appear

Reviewers: aria

Reviewed By: aria

Differential Revision: http://phabricator.khanacademy.org/D13431