Fix numeric-input articles with this one weird trick
Summary:
Use defaultProps on the numeric-input widget to make it not crash when adding
it to an article.

This is roughly because we render a default numeric-input before the first
prop change comes from the editor; so we need valid defaultProps on the
widget instead of just the widgetEditor.

Auditors: michelle