Fix perseus editor not having the right width
Summary:
The editor portion of the perseus editor didn't have the correct width.
It looks like a bad interaction between `table-layout: static`, `width`,
and `padding-right`. Adding `min-width` seems to force the width not to
shrink, producing the correct output. This makes it match with what we
see in the local perseus editor.

Test Plan:
 - Subperseus
 - Ensure that the editor now has a width of `360px` (`390px` with
   padding)

Reviewers: alpert, aria

Reviewed By: aria

Differential Revision: http://phabricator.khanacademy.org/D14289