Add widget ID to group annotation callback.
Summary:
The number is very useful in order to know in what order rendering
occurred, but getting any additional information on the widget is
difficult with just the number.

In the case that prompted this change, I needed to figure out the score
for the group widget, which is provided via a mapping from widget IDs
to score objects.

Test Plan:
Verified that arguments provided to the group annotation function were
reasonable and coincided with the correct widget.

Reviewers: aria, alex

Reviewed By: alex

Subscribers: #sat, eater, dylan

Differential Revision: http://phabricator.khanacademy.org/D14606