Fix some graphie stuff in the IEs
Summary: The `.remove()` function on DOM elements seems to not
exist (thanks IE), so in one instance it was replaced with the jQuery
equivalent, and in another place it was unnecessary, since
`.appendChild` automatically moves elements to the end.

Also, IE9 doesn't support 3d transforms, so the `translateZ(0px)` was
causing none of the styles to work. I added a function to check if 3d
transforms are allowed, and if not, doesn't add that to the transform.

Test plan:
 - Make a number line in IE 9-11, ensure that moving the point works
 - Add an interactive in IE 9-11, ensure that adding and removing a
   movable point work

Auditors: aria, charlie