Changes from testing with realish data.
The all/some interface didn't work with perseus, since it needs to
accept serializable values. Moving to "all"/"some".

Make a test case for "some".

Update lens to the latest version due to an error.

acceptedUnits used to have a type error - sometimes an array of string
and sometimes a string. Now it's always an array of string. Also add
proptypes.

Add a way to extract the base units from a unit with coefficient.

Auditors: alex

Test plan: make test