Fix fancy-dropdowns z-indexing
Summary:
The z-indexing of fancy dropdowns was broken, because the
`position: relative` of the `perseus-widget-dropdown` div was scoping
the z-index of the inner elements, so their `z-index: 4` wasn't
correctly placing them above other dropdowns. This moves stuff around in
order to accomplish this while still letting things be the correct
sizes. (Note that the widget is no longer `position: relative` when the
fancy dropdowns are used, but is still on webapp).

Test Plan:
 - Load the 'properties of shapes' exercise on mobile and do problems
   until you get one with two dropdowns on top of one another.
   - Click the top dropdown, ensure that it opens and covers up the
     bottom dropdown.
   - Ensure that the size of the options that drop down are the same
     size as the large dropdown.
 - Load the 'properties of shapes' exercise on webapp and do problems
   until you get one with two dropdowns on top of one another.
   - Enable the scratchpad, ensure that you cannot draw on top of the
     dropdowns.

Reviewers: aria

Reviewed By: aria

Differential Revision: https://phabricator.khanacademy.org/D15224