Remove the plotter's use of translate
Summary:
It looks like when the plotter was created, it added some code
to attempt to align the top of the selection bars to the lines that were
supposed to be selected. However, now it looks like that code doesn't
really work; when the bar itself is moved, the bar aligns its top to the
line, but when the mouse is taken away, the bar moves back and is
centered on the line. To do this, the plotter used a `translate`
function which isn't used anywhere else in perseus, and is no longer
working because of the wrapped-line changes. So, I'm going to remove
this functionality and maybe in the future we can make it work better.

Test Plan:
 - Visit an item with a plotter, like
   [this](http://perseus.ka.local/test.html#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20plotter%201%5D%5D%5Cn%5Cn%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22plotter%201%22%3A%7B%22type%22%3A%22plotter%22%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22correct%22%3A%5B1%2C1%2C1%2C1%5D%2C%22starting%22%3A%5B1%2C1%2C1%2C1%5D%2C%22type%22%3A%22bar%22%2C%22labels%22%3A%5B%22%22%2C%22%22%5D%2C%22categories%22%3A%5B%22x%22%2C%22y%22%2C%22z%22%2C%22w%22%5D%2C%22scaleY%22%3A1%2C%22maxY%22%3A10%2C%22snapsPerLine%22%3A2%2C%22labelInterval%22%3A1%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22type%22%3A%22multiple%22%2C%22options%22%3A%7B%22content%22%3A%22%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%7D%7D%2C%22calculator%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D)
   - Make sure that there are no javascript errors on loading the item.
   - Make sure that moving the line causes no errors.

Reviewers: eater

Reviewed By: eater

Subscribers: aria, alex

Differential Revision: https://phabricator.khanacademy.org/D15681