Expression: Prevent typing '*' from breaking you out of a denominator
This happened when the multiplication sign in an expression widget
was set to display as \times, and the keyboard was used to enter a
multiplication sign. This ended up calling `MathField.latex(<latex>)`,
which does the opposite of an incremental update: it blows everything
away, then recreates it from scratch. In the process, the cursor
position is reset to the end of the input.

This is okay in certain cases, such as when typing '<>' to insert ≠, but
very frustrating in the common case of typing '*' while in a denominator
(or other non-end section of the input).

The commit detects this common case and handles it separately, by
deleting the freshly inserted \cdot and replacing it with \times.

No action is required for the opposite case (multiplication displayed as
\cdot) because that is the default action for typing '*'.

Test Plan:
Go to
http://localhost:9000/test.html#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20expression%201%5D%5D%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22expression%201%22%3A%7B%22type%22%3A%22expression%22%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22answerForms%22%3A%5B%7B%22value%22%3A%22%22%2C%22form%22%3Afalse%2C%22simplify%22%3Afalse%2C%22considered%22%3A%22correct%22%2C%22key%22%3A0%7D%5D%2C%22buttonSets%22%3A%5B%22basic%22%5D%2C%22functions%22%3A%5B%22f%22%2C%22g%22%2C%22h%22%5D%2C%22times%22%3Atrue%7D%2C%22version%22%3A%7B%22major%22%3A1%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22type%22%3A%22multiple%22%2C%22options%22%3A%7B%22content%22%3A%22%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%7D%7D%2C%22calculator%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D
Type '1' '/' '2' '*' '3' and see `\frac{1}{2 \times 3}`

Auditors: aria