[number-line] NumberLineEditor cleanup!
Summary:
Cleaned up the number-line editor a bit. The main goal was
to remove uses of the `NumberInput`'s `label` prop, but
I also cleaned up a lot of capitalization and colon
consistency.

The position RangeInput still has weird left-margin which
makes me sad, but that is for another day.

Also tickstep is broken if you type in "0.x", but that is
also for another day.

Test Plan:
1. Made a number-line and looked at the editor.
2. Edited all the options.
3. Saw that tick step broke when I tried to type `0.2`
4. Tried grading and verified that it worked for empty/wrong/correct.
5. Checked that tick step was broken in master (it is).
6. Was sad.

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D16307