Remove non-SAT clues after all
Summary:
There are two high level ways to see clues: in SAT and outside of SAT (but within an AB test). Since the AB test was inconclusive, I had opted for showing clues outside of SAT to be consistent with SAT. Here are the related diffs that made that happen:

- webapp https://phabricator.khanacademy.org/D16411
- perseus https://phabricator.khanacademy.org/D16412
- khan-exercises https://phabricator.khanacademy.org/D16413

Unfortunately, enabling clues outside of SAT caused some problems for clues *inside* SAT. Because of the frequent `render.scoreWidgets()` calls in sat-perseus-renderer.jsx, we keep calling simpleValidate, which in turn would set `showClues` to true. We would then show the rationale for an answer even before the user had clicked "Check answer". This wasn't a problem when clues were hard-coded to be enabled for 5 non-SAT questions, but now it is.

Ultimately, I'd like to fix this by understanding why we need to set `showClues` to true in `simpleValidate`. It seems like we should be able to "validate" something (check an answer's correctness, yes?) without necessarily mutating this other state of "should we show clues?". Maybe we can set the clues visibility in some sort of "interact with the check answer button" function, but I don't have enough familiarity or time with perseus to pursue that avenue now. I'll add a TODO to leave a trace for a future Marcia.

Test Plan:
Visit a perseus exercise that has clues content:
Visit /exercise/naming-shapes-2
Get the answer wrong
See no clues pop up
Get it right
Proceed to next question

Visit something in /mission/sat (after passing the show_sat_mvp bridge)
Click on an answer (don't check answer yet)
See no rationale pop up (it did before, all excitedly and at the wrong time)
Click check answer
Now see a rationale appear

Reviewers: eater, aria

Reviewed By: aria

Subscribers: emily, aria, #sat

Differential Revision: https://phabricator.khanacademy.org/D16445