[passage-markdown] Fix circleLabel parsing
Summary:
circleLabels didn't have an `order` field, and were not
passing `state` to the recursive parse call.

This commit fixes those bugs and adds a test case for them.

This commit also fixes `squareLabel` not having an `order`,
but does not add a test case for it (since it happened
to be working :( ).

Test Plan:

    make test

Auditors: emily, marcia