Upgrade to React 0.13 (RC2) and bump major version
Summary: I fixed a lot of key warnings as we now consider these must-fix and a handful of other things.

Test Plan: Use each widget from the dropdown and click around the editor to try to trigger most code paths. The table widget threw some errors that I think are unrelated to this; everything else worked.

Reviewers: aria

Reviewed By: aria

Differential Revision: https://phabricator.khanacademy.org/D16382