[radio] remove .perseus-interactive from <label>
Summary:
.perseus-interactive is already specified on the option content,
so this class should be unnecessary. More importantly, this class
was causing styling problems in SAT, because they have
`position: absolute` specified on the <span.checkbox>

Test Plan:
`make subperseus` and go to /mission/sat, and see that the
(A) (B) (C) (D) (E) circles are centered in the multiple-
choice choices.

Auditors: marcia