Marking particular widgets as accessible or not
Summary:
Ref T3702

We'll be tagging AssessmentItems are accessible or not, depending on the widgets
they contain, and an override switch by the content creator. This changeset includes
tagging individual widgets as being accessible or not, and also contains a lookup
method in the Widgets class.

I've tagged some current widgets are accessible - meaning I can use them without issue
with a screen reader and keyboard.

Test Plan: . `make test`

Reviewers: emily, john, aria

Reviewed By: aria

Subscribers: #a11y

Maniphest Tasks: T3702

Differential Revision: https://phabricator.khanacademy.org/D17137