Fix SAT text selection bug in Safari.
In Safari, if a user clicked an option in a radio widget, all the text
in that radio widget would be selected (usually, but not always). This
fixes that behavior (but the text is still selectable as normal).

I do not know why this fixes the behavior, but I am fairly sure it's
just working around some bug in Safari.

Test Plan:
1) Go to /mission/sat in Safari
2) Open up a task
3) Click on some multiple choice options (make sure to click the text
   part of the option, rather than the blank space). See that the
   entire text is not selected.

Auditors: marcia