Fix strange background glitch.
Summary:
When in two column mode, you can get into a fairly strange state where
the background of the left-column div is not being applied.

{F128536}

This is an odd bug and probably has its root in some undefined behavior
or some browser bug. See the Test Plan for the very unusual
reproduction steps.

Test Plan:
In sat-mission.less in the web-app repo, change `&.showing-passage {`
to `&.showing-passage .content-area {` (this is a companion change to
this that I will audit into webapp before brining in this change).

Using Chrome...

1) Go to /mission/sat
2) Open the first diagnostic quiz
3) Find a question with two columns, but it all fits onto your screen
without creating scroll bars.
4) Open up the dev console at the bottom of the page such that it
forces scroll bars to appear.
5) Pop out the dev console so that the page goes back to fitting onto
your screen.
6) See that the background does not appear like in the screenshot.

Reviewers: marcia

Reviewed By: marcia

Subscribers: aria, #sat

Differential Revision: https://phabricator.khanacademy.org/D17424