Wrap radio buttons in a fieldset and provide a helpful legend for screen reader users. Fixes T3711.
Summary: This change makes it so that radio buttons are now wrapped in a (hidden) fieldset and have a helpful legend that provides an explanation of what the user is expected to do.

Test Plan:
Navigate to a radio list using a screen reader. It should read off the helpful legend. Example:

https://www.dropbox.com/s/7olag5g1jd18gci/Screenshot%202015-04-24%2010.33.17.png?dl=0

Reviewers: johnsullivan, jordan, aria

Reviewed By: jordan, aria

Subscribers: #a11y, #sat

Maniphest Tasks: T3711

Differential Revision: https://phabricator.khanacademy.org/D17526