Add additional text to make it obvious if a selection is selected, correct, or incorrect. Fixes T3725.
Summary: This adds some additional text to the radio choices to make it obvious which options have been checked and if it's marked correct or incorrect.

Test Plan:
Use a screen reader to navigate the radio options in a Perseus problem. When you select an option it should now say "Checked" and when it's correct or incorrect it should say "Correct" or "Incorrect".

A gif demo of it in action:
{F130850}

(Perseus doesn't seem to have the rubric populated, at least in my testing in the testrender.html, I'm optimistic that it'll work once run in the KA site.)

Reviewers: aria, jordan, johnsullivan

Reviewed By: aria, jordan, johnsullivan

Subscribers: #sat, #a11y

Maniphest Tasks: T3725

Differential Revision: https://phabricator.khanacademy.org/D17530