Inform the user about missed correct answers. Fixes T3738.
Summary: We need to make sure that when the user answers a question incorrectly that there is some indicator as to which answer was the correct one. We need to bring in some text to that effect.

Test Plan:
Make sure that a correct answer that is un-checked now has hidden text that indicates that it was the correct answer:
https://www.dropbox.com/s/knlmlrsd0mdan53/Screenshot%202015-04-29%2010.54.02.png?dl=0

Reviewers: johnsullivan, aria

Reviewed By: johnsullivan, aria

Subscribers: #sat, #a11y

Maniphest Tasks: T3738

Differential Revision: https://phabricator.khanacademy.org/D17599