Make a built version of perseus that can be required
Summary:
Make a (scary) version of perseus that can be required
externally in node and used to do things with. In particular, this might
be used to correctly parse item data for i18n purposes, or to run
exercise backfills that actually understand perseus.

This required some scariness that is mostly in node/require.js, which is
an (almost) duplicate of node/environment.js with some stuff
removed.

I also added some tests to make sure that things work-ish, and added the
building and copying to the Makefile build process.

Depends on D17372

Test Plan:
 - See that `make subperseus` works (which both builds it and runs the
   tests).

Reviewers: aria

Reviewed By: aria

Subscribers: alpert, eater, alex

Differential Revision: https://phabricator.khanacademy.org/D17379