Fix hints/focus problem when proceeding to next q
Summary:
Before this changeset, every time the hint # changed, we would try to focus on a hint. This is generally okay as we show more hints, but it's not okay when we're proceeding from one question (with hints shown) to the next question (with no hints). There are no hints, after all, to be focused on! This causes js errors and prevents the next question (and thus the task) from being completed.

With this changeset, we just try to focus when the hint # increases. I suspect this changeset does the job adequately -- we only try to focus when the # gets larger. Of course one could be even more defensive and check that that dom element is there, too, but I'm not sure if that's necessary?

Test Plan:
Pass show_sat_mvp_private_beta gandalf bridge
Visit /mission/sat
Open up a math practice task (one of the list of skills further down the page)
Take a few hints - see that focus is being given to each new hint
Answer the question
Click Check button
Click Next button

Verify that you can see the next question!

Reviewers: johnsullivan, aria

Reviewed By: aria

Subscribers: #a11y, aria, #sat

Differential Revision: https://phabricator.khanacademy.org/D17982