getUserInput: Fix up large `getUserInput`s
Summary:
`getUserInput` should return the actual user input to a widget,
not "all the props on the widget ever" and especially not
"all the props on the widget including the default ones like
enabledFeatures and apiOptions". This commit addresses three
problematic `getUserInput` functions (two of which were
`WidgetJsonifyDeprecated` mixins), which prevents an
unnecessarily large `/attempt` call on [SAT passage items][1],
preventing a 500 when attempting them in exercise mode.

This might have good ramifications for other exercises, if anything
else was getting ridiculously long responses due to radios,
which were a significant source of the length.

[1]: http://localhost:8080/v1-sat/v1-sat-reading-and-writing/v1-reading/v1-reading-science/e/v1-reading-science-1

Test Plan:
1. subperseus
2. Do the SAT item linked in the summary
3. Click "check answer"
4. Don't get a red "refresh" error bar.

Auditors: eater