Adds video widget.
Summary:
The video widget is an impossibly easy-to-use widget for embedding a video in an article. You just give it the URL and you're all set.

You can paste in a YouTube video URL or a KA embed URL and either will be normalized to the KA embed URL, which ensures that video watching points are recorded. This widget does also support embedding videos from other sites (e.g. Vimeo), but no points will be tracked.

This widget assumes the presence of D19321 to remove the extra controls from embedded videos.

Videos default to "block" alignment but can also be aligned full-width and float-left/right. The float sizes will be very small, but could still be reasonable for videos that are focused on a person rather than drawings. Plus, the user can go full screen with those videos.

Test Plan:
If you see tota11y in the embedded video area, you can turn that off by adding &tota11y=0 to the URL. Non-developer users would not see that.

1. Preview a video added by URL in [[http://localhost:9000/article.html#content=%5B%7B%22content%22%3A%22%5B%5B%E2%98%83%20video%201%5D%5D%5Cn%5CnHi%20I%27m%20a%20particle%21%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22video%201%22%3A%7B%22type%22%3A%22video%22%2C%22alignment%22%3A%22block%22%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22location%22%3A%22https%3A%2F%2Fwww.youtube.com%2Fembed%2FJC82Il2cjqA%22%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%5D| block alignment ]].
2. Preview a video in [[http://localhost:9000/article.html#content=%5B%7B%22content%22%3A%22%5B%5B%E2%98%83%20video%201%5D%5D%5Cn%5CnHi%20I%27m%20a%20particle%21%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22video%201%22%3A%7B%22type%22%3A%22video%22%2C%22alignment%22%3A%22float-left%22%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22location%22%3A%22https%3A%2F%2Fwww.youtube.com%2Fembed%2FJC82Il2cjqA%22%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%5D| float-left ]].
3. Preview a video in [[http://localhost:9000/article.html#content=%5B%7B%22content%22%3A%22%5B%5B%E2%98%83%20video%201%5D%5D%5Cn%5CnHi%20I%27m%20a%20particle%21%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22video%201%22%3A%7B%22type%22%3A%22video%22%2C%22alignment%22%3A%22full-width%22%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22location%22%3A%22https%3A%2F%2Fwww.youtube.com%2Fembed%2FJC82Il2cjqA%22%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%5D | full-width ]].
4. [[http://localhost:9000/article.html|Create a new article]] and paste https://www.khanacademy.org/humanities/music/music-basics-notes-rhythm-music/v/lesson-1-note-values-duration-and-time-signatures as the location. You should see the field reduced to just `lesson-1-note-values-duration-and-time-signatures`. The video should show up in the preview.
5. Watch some of the video and verify on your dashboard that your points have increased.

Reviewers: alex

Reviewed By: alex

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D19037