[Static Widgets 3] Use staticTransform to render widget state
Summary:
This diff hooks up the `staticTransform` method for widgets with the
Renderer to actually render initial state. This means that static
widgets work sorta!

Specifically, widgets that export a `staticTransform` method show a
toggle to render as static or not. If static mode is set, the
`staticTransform` method is called to extract out the data needed for
the initial render of the widget.

Here's what it looks like in action:
{F165319 size=full}

Next steps:
- Overlay a div on top of widgets in static mode so they
  can't be interacted with, and think a more robust solution for the future.
- Implement the `staticTransform` method for our widgets.

Test Plan:
Visit this Perseus link.
- Verify that the categorizer widget's answer gets transferred over to
  the rendered widget when static mode is on.
- Verify that changing the widget's answer also propogates changes to the
  rendered widget.
- Verify that turning static mode off resets the rendered widget state.

Reviewers: aria

Subscribers:

Differential Revision: https://phabricator.khanacademy.org/D20169