Update react-components to include katex-a11y
Summary:
For https://app.asana.com/0/27216215224639/45365017532302

John made this katex-a11y thing that uses katex's parsing to turn math into
plain language suitable for screen readers. It's deployed on the site but not
working for any Perseus things because Perseus's react-components is older than
webapp's and we bundle it with Perseus rather than using webapp's.

Test Plan:
make subperseus
Load /mission/sat in webapp and find my way to something with math in it
Inspect the rendered math and find an sr-only <span> nearby with "start fraction, 3, divided by, 7, end fraction" in it.

Reviewers: emily, aria

Reviewed By: aria

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D20455