Adds perseus.sublime-project to the project.
Summary:
There are some directories that are handy for Sublime to ignore, so I've
added a perseus.sublime-project file to the repo so that you can open that
and have them ignore automatically. We'll also be able to put project-specific
settings in there.

Also adds the sublime-workspace file to .gitignore.

Test Plan: - Open perseus.sublime-project in Sublime

Reviewers: alex, aria

Reviewed By: aria

Subscribers: aria

Differential Revision: https://phabricator.khanacademy.org/D19480