JIPT: Perseus article paragraph jipt translating
Summary:
Ohmygoshhowcaniexplaineverything

Exercises have entire renderers translated. Articles, however, are
translated paragraph-by-paragraph. This means instead of attaching
crowdin jipt to the renderer as a whole, we need to convince crowdin to
attach it to each paragraph.

To achieve this, we cast several renderer spells, a couple markdown
inscriptions, and invoke some crowdin chants, and hope our arcany is
correct.

Basically, we render the crowdin IDs in the outer QuestionParagraphs
when we find them, and we make sure to render the same QuestionParagraph
structure.

Depends on D21244

Test Plan:
 * `make subperseus`
 * go to [a particle's jipt page][1]
 * edit a paragraph
    * try changing the type (paragraph -> list; paragraph -> code)
    * try turning one paragraph into two by adding '\n\n's

[1]: http://localhost:8080/science/biology/chemistry--of-life/elements-and-atoms/a/atomic-number-atomic-mass-and-isotopes-article?lang=en-pt

Auditors: emily, jlfwong, kevinb, james

Previous Differential Revision: https://phabricator.khanacademy.org/D21245
(sent as an audit)