Improve Perseus radio accessibility
Summary:
https://app.asana.com/0/45794228513726/45365017532278

Background:
-----------

We wanted a widget with radio-like behavior, but to be able to
style it such that the native radio control was hidden for the SAT mission.
But, we ran into accessibility trouble: an unrendered radio button can't have
focus. So, we had to apply `tabindex` and `role="radio"` to the wrapper
elements, and forward focus events on them to the underlying radio. This made
the widget usable with a keyboard-only interface, but not in the conventional
way; you could only select elements via tab+space, rather than the standard
up-down arrows.

Solution:
---------

Rather than try to reimplement native behavior, let's just go back to regular
radio buttons. By applying the correct styles (specifically,
`.ui-helper-hidden-accessible` borrowed from webapp), we can make the radio
button //look// exactly like `display: none` would look, but still be
physically present on the page and able to receive focus. This allows us to
remove a nontrivial amount of hax to make a `li` act like a radio button :)

Since we're removing focusability from the wrapper element, we need to change
our HTML and CSS accordingly. Now, the radio button is the element with focus,
and, because of the set of CSS selectors available, you can only style an
element's self, siblings, or descendants according to that element's `:focus`
state. Therefore, instead of applying borders, backgrounds, etc. to the
wrapper `li`, we introduce a new `.description` element as a sibling of the
radio `input`. This is the element that actually gets styled in the SAT
mission, and contains all of the visible content of the `li`.

Test Plan:
----------

* Try radio buttons in test mode.
  * Confirm no visual regressions in single-select or multi-select mode.
  * Confirm that a screen reader says reasonable things.
* Try radio buttons in the SAT mission.
  * Set up.
    * `make subperseus`
    * Visit http://localhost:8080/mission/sat/practice/math
    * Choose any task containing multiple choice questions.
  * Choose an answer with the keyboard.
    * Confirm that keyboard controls function exactly as they do with native radio buttons.
    * Confirm no visual regressions.
    * Confirm that focused state, selected state focused+selected state, and active state are all clearly visually distinct.
    * Confirm that the screen reader says helpful, accurate things along the way.
  * Choose an answer with the mouse.
    * Confirm that mouse controls function exactly as they do with native radio buttons.
    * Confirm no visual regressions.
    * Confirm that focused state, selected state focused+selected state, and active state are all clearly visually distinct.
  * Submit an incorrect answer.
    * Confirm no visual regressions.
  * Submit a correct answer.
    * Confirm no visual regressions.

Reviewers: john, johnsullivan

Reviewed By: johnsullivan

Differential Revision: https://phabricator.khanacademy.org/D21349