Molecule rendering, part 2: layout engine
Summary:
For the hackathon, I made some perseus widgets that can render simple 2D
molecule structure drawings into perseus content.  For the general approach,
see part 1 (D21679).

This revision is the second part of the rendering pipeline: the layout engine.
This portion takes the molecular parse tree as input and outputs a series of
rendering instructions, which are things like:
`{type: "line:single", startPos: [<x>, <y>], endPos: [<x>, <y>]}`
that say what to render and where, but not how to do it.  (The goal in doing
this is to decouple as much as possible the computation of what needs to be
rendered from the renderer itself, so that we could, e.g., swap out an html5
canvas for a native mobile view.)

Depends on D21679

Test Plan: `make test`

Reviewers: alex

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D21680