Molecule rendering, part 3: canvas renderer and molecule widget
Summary:
For the hackathon, I made some perseus widgets that can render simple 2D
molecule structure drawings into perseus content.  For the general approach,
see part 1 (TODO: diff number).

This revision is the the third part of the rendering pipeline, the canvas
renderer, along with a basic molecule widget that wraps the canvas element.
The canvas renderer takes a 2D canvas drawing context as well as the rendering
instructions that were the output of part 2 (TODO: diff number), and it renders
the instructions into the drawing context.

The molecule widget wraps the canvas element and orchestrates the whole
rendering pipeline, from SMILES string to rendered canvas.  It also allows
rotation of the entire molecule, since the orientation in which it's drawn is
somewhat arbitrary.  Since the molecules we can reasonably render as a 2D
diagram are generally pretty small, there's no problem doing this whole thing
in real-time.  Thus, the only content that we need to store for the widget is
the SMILES string itself and the rotation angle.

Depends on D21680

Test Plan:
- in the perseus devserver, add a molecule with smiles CC(=O)CC#CCCCC(Cl)(Cl)Cl
  and rotation 180 degrees, or visit:
```
/testrender.html#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20molecule-renderer%201%5D%5D%5Cn%5Cn%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22molecule-renderer%201%22%3A%7B%22type%22%3A%22molecule-renderer%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22static%22%3Afalse%2C%22smiles%22%3A%22CC%28%3DO%29CC%23CCCCC%28Cl%29%28Cl%29Cl%22%2C%22rotationAngle%22%3A180%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22type%22%3A%22multiple%22%2C%22options%22%3A%7B%22content%22%3A%22%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%7D%7D%2C%22calculator%22%3Afalse%2C%22periodicTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D
```
(which is a molecule that hits all the major rendering functionality)
- verify that it renders as in the screenshot

Reviewers: alex

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D21683