Update babel options to match webapp and add babel polyfills
Summary:
We'd like perseus to have access to the same ES6/7 language features
and polyfills as webapp so that support is consistent across the two
code bases.  I've made a small refactor that rewrites _.each to use
`for of` and `Object.entries` to verify that things are working correct
with the new language features an polyfills.  I've also added the same
sanity checks that webapp has to verify that the polyfill is working
correctly.

Test Plan:
- make test
- load test.html, article.html and index.html and verify that you can copy and paste widgets.
- make subperseus
- edit an exercise http://localhost:8080/math/algebra/sequences/constructing-geometric-sequences/e/recursive-formulas-for-geometric-sequences/edit/xa591ffeb53a30914 and verify that you can copy and paste widgets.

Reviewers: john, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D22580