trigger onRender when an image loads
Summary:
https://app.asana.com/0/27216215224639/79823010109644

When a Matcher widget renders, it measures its contents in order to
align them correctly. However, when the Matcher contains images,
those images are initially small loading icons, then later grow into
full-sized images, without ever being re-measured.

This change adds an `onUpdate` callback to `SvgImage` and `ImageLoader`,
which triggers when the image's DOM updates from the spinner to the
full-size image. The image's `Renderer` watches for this event,
and propagates it as `onRender` to the parent — which, in the case of a
`Matcher` parent, will cause it to re-measure its children and re-align
them. Hooray!

Test Plan:
1. Before making this patch, visit http://localhost:8080/science/biology/chemistry--of-life/skill-check-chemistry-of-life/e/chemical-bonds-and-reactions/
2. Observe that this exercise renders incorrectly. {F245244}
3. Apply this patch to webapp via `make subperseus`.
4. Visit http://localhost:8080/science/biology/chemistry--of-life/skill-check-chemistry-of-life/e/chemical-bonds-and-reactions/
5. Observe that the previously-broken exercise now renders correctly. {F245245}

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D24948