Add zero-rating-friendly link parser
Summary:
Similar to the url-sanitization that's done in SimpleMarkdown's base link parser, we're adding a zero-rating-friendliness extension to that parser to fix zero-rating-unfriendly links for zero-rated requests. We're doing this on the client-side to avoid any accidental (and undesirable) cacheing of zero-rated links that could happen if we did this transformation server-side. This logic is based loosely off the logic in webapp/javascript/shared-package/absolute-links.js, but gets to be a little simpler because we know that links won't already be zero-rated.

This is the Perseus equivalent of https://phabricator.khanacademy.org/D26187.

Test Plan:
* Run `make subperseus` to build Perseus
* Enable zero-rating mode by visiting `localhost:8080/zero/enable` and `localhost:8080/zero/allow`
* Visit a Perseus article (e.g. `http://localhost:8080/science/biology/chemistry--of-life/electron-shells-and-orbitals/a/the-periodic-table-electron-shells-and-orbitals-article`)
  * Load the editor and edit the article to ensure that there are both interal and external links.
  * Ensure that internal links are rewritten to point at zero.khanacademy.org, and external links are rewritten to point at the `/zero/external-link` warning interstitial
* Disable zero-rating mode by visiting `localhost:8080/zero/disable`
* Visit the same Perseus article and ensure that internal and external links have not been rewritten

Reviewers: john, emily

Subscribers: benkraft, jared

Differential Revision: https://phabricator.khanacademy.org/D26197