Add overflow:hidden to graphie containers
Summary:
On small screens, the graphics are scaling fine but the svg remains the same
size, rendering parts of the page unclickable.
See asana https://app.asana.com/0/99293390018551/111130038595836

@emily & @alex -- can you think of a reason that it would be a problem to do
this?

{F457087}

Test Plan: Subperseus, then go to [[ http://localhost:8080/math/geometry/similarity/intro-to-triangle-similarity/e/similar_triangles_1 | this exercise ]] on a small screen, and verify that the "stuck? take a hint" button is clickable.

Reviewers: alex, kevinb, emily

Reviewed By: kevinb, emily

Subscribers: emily, alex, #perseus

Differential Revision: https://phabricator.khanacademy.org/D26532