Only use jsx-loader on src/ and react-components
The purpose of this is to prepare for adding jsx-loader processing of .js files as well (for Babel). This commit changes essentially nothing because all the .jsx files used in this project are already in those directories.

The only .jsx files not in those directories are test files we don't use under:
./node_modules/i18n-babel-plugin/test/fixtures/
./node_modules/jest-cli/src/__tests__/test_root/

Those were found by running:
find . -name \*.jsx | grep -v -e '^\./src' -e '^\./react-components'