Remove ESLint exceptions for 'react/prop-types'
I took a conservative approach to deciding whether to mark props as required or not. I only marked a prop as required if it was very obvious that it should be, for example because it would make render() throw if it was missing, or because it was an onChange function prop on a widget editor.