Change all vars to consts
This is an intermediate step as part of removing exceptions for the 'no-var' ESLint rule. This particular commit will not produce a working build because obviously many of those vars should actually be lets rather than consts.

I was a bit surprised that a tool doesn't seem to exist for doing this properly using a parser, and I was tempted to write one, but in the end I did it manually:

I grepped for `(^|\s)var(\s)`, inspected the output to make sure that there weren't any occurences inside strings or anywhere else we wouldn't want to replace them, and then replaced the regex with `\1const\2` using Sublime Text.