Fix lint in Perseus
Summary: Perseus is going to be a real submodule now, so it can't have
any lint in it any more. This fixes alllll the lint, including:

 - Adding submodules and stylesheets to the lint blacklist, because
   we're not going to fix the lint there.
 - Fixing a bunch of `var` -> `const` errors in miscellaneous files.
 - Updating the khan-exercises submodule to a commit that doesn't have
   lint in it.
 - Fixing the lint in react-components and substating that

I also made it so that `make build` runs the linter. I'm not 100% sure
this is kosher though, since it depends on the khan-linter stuff? I'd
like to have travis run the lint tests also but I'd have to install the
linter on travis for it to work.

Test Plan:
 - `ka-lint`
 - `make build`

Auditors: alex, csilvers