Add keypad configuration logic to `Expression` widget
Summary:
This diff adds the logic to determine which version of the Expression keypad ('Basic' or 'Advanced')
to show, based on the selected button sets. Right now, there's no concern for scraping the necessary
variables and constants from the answer area--that will come later.

Note that the `keypadConfiguration` is computed in the `editorProps` transform to keep the component
itself dumb, and accommodate for the fact that the `keypadConfiguration` will be a function of the
_answer_ to the problem, which is a property of the editor/grader and not of the actual widget.

On Asana @ https://app.asana.com/0/114493820703878/132295112211115.

Test Plan:
- Since the config logic isn't hooked up to the keypad yet, add `console.log(this.props.keypadConfiguration)` to the `render` method of `Expression`.
- Open up an `Expression` widget.
    - Verify that the keypad type is `BASIC_EXPRESSION`.
- Enable the 'trig' button set.
    - Verify that the keypad type switches to `ADVANCED_EXPRESSION`.
- Enable the 'prealgebra' button set.
    - Verify that the keypad type stays as `ADVANCED_EXPRESSION`.
- Disable the 'trig' button set.
    - Verify that the keypad type stays as `ADVANCED_EXPRESSION`.
- Disable 'prealgebra' and enable some other button type.
    - Verify that the keypad type switches to `BASIC_EXPRESSION`.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27270