Display/hide keypad on input focus/blur
Summary:
This is the initial integration between math-input and perseus.
There is still lots of work to do.

This diff in particular handles focus/blur of single expression
widgets on the page.  I a user focuses a different widget type
it will hide the keypad.

Depends on https://phabricator.khanacademy.org/D27220.

Test Plan:
- load localhost:9000/test.html (perseus' test app)
- add an expression widget
- click to focus, see keypad come up
- type some numbers on the keypad
- click outside of the input and keypad, see keypad goes down

Reviewers: emily, charlie

Reviewed By: charlie

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27222