Add keypad configuration logic for non-`Expression` inputs
Summary:
In D27270, we added the keypad configuration logic for the `Expression` widget. Here, we do the
same for `InputNumber` and `NumericInput`, the other input-enabled widgets.

The logic in these two cases is actually somewhat more complex. They both attempt to use the
simplest keypad available ('Number' when possible, often 'Fraction', but sometimes
'Basic Expression' with Pi), but their capabilities in doing so are somewhat different, as the
`NumericInput` automatically sets its answer format when a content creator includes Pi, and the
`InputNumber` does not, while both store just the resolved value of their answers and not plaintext.

On Asana @ https://app.asana.com/0/114493820703878/132295112211117 and https://app.asana.com/0/114493820703878/132295112211119.

Test Plan:
- Open up a `NumericInput`.
    - Add some answers with non-integer input.
    - Open up the renderer.
        - Verify that you get the Fraction keypad.
    - Add an answer with Pi in it.
    - Open up the renderer.
        - Verify that you get the Basic Expression keypad, with Pi available.
    - Confine your answers to integer input.
    - Open up the renderer.
        - Verify that you get the Number keypad.
    - Verify that grading works as expected.
- Open up an `InputNumber`.
    - Use non-integer input in the answer (fraction, decimal, etc.) without setting the answer type.
    - Open up the renderer.
        - Verify that you get the Basic Expression keypad, with Pi available.
    - Use integer-only input in the answer without setting the answer type.
    - Open up the renderer.
        - Verify that you get the Number keypad.
    - Verify that setting an explicit answer type gives you the expected keypad (Basic Expression for "pi", Number for "integer", and Fraction for anything else).
    - Verify that grading works as expected.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27291