Remove the i18n-babel-plugin
Summary:
We'd like to upgrade to Babel 6 so that babel knows how to deal
with .babelrc files in submodules such as math-input which is
already using Babel 6.  We're require math-input from source b/c
we don't want to include it's version of React.  This diff is
the first step in the process.

In order to upgrade to Babel 6 we need to get rid of the i18n-babel-plugin
b/c it only works with Babel 5.  In order to do that we convert
all instances of <$_> to $_() calls.  There were no instances of
<i18nDoNotTranslate> so we don't have to worry about converting
those.

Actually upgrade to Babel 6 will appear in a following diff.

Test Plan:
- make test
- make build

Reviewers: emily, charlie

Reviewed By: emily, charlie

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27350