Substate `math-input` and rebuild to pass webapp's JS tests
Summary:
`math-input` was using String iteration, which is not polyfilled in webapp and was causing some JS
tests to fail. This was fixed in D27352 and is included in this substate, along with a re-build and
a modification to the Makefile to make building `math-input` a little more straightforward.

Test Plan:
- From webapp:
    - Run `tools/runjstests.py javascript/manticore-package/`.
    - Run `tools/runjstests.py javascript/perseus-package/`.
    - Run `tools/runjstests.py javascript/perseus-admin-package/`.
- Verify that all tests pass (previously, these packages contained failing tests).

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27353