Define Node environment for modules required by webpack
Summary:
Redux was throwing a warning (that webapp's logging was treating as an error) because the code
was minified, but the Node environment wasn't set to `production`. It seems that Webpack requires
that you explicitly provide the Node environment (e.g., with the `DefinePlugin` that I've used here)
for `require`d modules--it won't read the real `process.env.NODE_ENV`.

Test Plan:
- Add `console.log(process.env.NODE_ENV === 'production')` to `math-input/node_modules/redux/lib/index.js`.
- Set `export NODE_ENV=production` in the terminal.
- Run `make serve`, open up an input rendered with the new keypad.
    - Verify that the console prints `true`.
- Set `export NODE_ENV=development`.
- Repeat.
    - Verify that the console prints `false`.
- (Trying to figure out how to test more confidently in webapp.)

Reviewers: kevinb, emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27355