Use compiled .js files from math-input
Summary:
Summary
https://phabricator.khanacademy.org/D27356 added a postinstall
script to math-input which generates .js files in lib/ that
have already been compiled.  This makes it much easier for
perseus to require them.  This diff makes a few minor changes
to account for the fact that we no longer need to compile these
files.

Test Plan: - make build

Reviewers: charlie, emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27357