Never use the Number keypad
Summary:
We've decided not to use the Number keypad and, instead, to always use the Fraction keypad for
the `NumericInput` and `InputNumber` widgets.

There are some concerns around the fact that these widgets are used in a few places on the site to
input Pi, which isn't included in the Fraction keypad. However, Cam has volunteered to lead the
charge on migrating those pieces of content over to the Expression widget by the time we ship the
beta.

Test Plan:
- Add a bunch of `NumericInput` and `InputNumber` widgets with various answers and settings.
- Open them in the renderer demo page.
- Tap on them all.
    - Verify that they all use the Fraction keypad, unequivocally.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb, jared

Differential Revision: https://phabricator.khanacademy.org/D27495