Update exercise color palette
Summary:
This is a drop-in change based on the file that @tabitha provided in the linked Asana ticket below.
The new color palette uses KA Green instead of the previous orange tone to indicate interactivity,
and updates many of the other colors to match Bibliotron.

On Asana @ https://app.asana.com/0/132409062806753/27549926164679.

Test Plan:
Open up a bunch of exercises; verify that they use orange instead of green.

{F542222}

Reviewers: david, benkomalo

Reviewed By: benkomalo

Subscribers: tabitha, jared

Differential Revision: https://phabricator.khanacademy.org/D27731