Revert `Explanation` widget to use class-based styles
Summary:
It's a pain to use inline styles for the `Explanation` widget's link styling without converting
more of Perseus to use inline styles. In particular, after the change to use inline styles, it's
challenging to style the font size responsively without using a CSS selector that makes stronger
assumptions as to the structure of the DOM.

Test Plan: Verify that the styles are as before for the `?article` page, with the XOM flag enabled (responsive) and disabled (16px).

Reviewers: benkomalo

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27773