Made numeric-input.jsx better recognize when a tooltip is needed
Summary:
There was a bug where a problem had both an accepted and incorrect answer with accepted answer types, the tool tip informing the user of the acceptable types was not displayed: https://app.asana.com/0/29625948558274/134275270870464

Why this happened: In the numeric input, when all answer forms are acceptable, the component will not display a tooltip (as all input is ok). To check this, it counts the number of acceptable input types, and if this is >= the total number of input types, it determines no tooltip is needed. However, there can be duplicate acceptable input types if some allow simplification (e.g. in the acceptable answer) and others don't (e.g. in the bad answer), overcounting the number of acceptable types.

To fix this, I count the unique input type names, instead of the total types.

Note: if for some reason the content creator has multiple answers which together cover all input data types, no tool tip will show. A more long-term fix might want to investigate why we even have input types for incorrect answers.

Test Plan:
- Verified in a browser that the tool-tip still displays when incorrect answers have additional valid answer types (**and now should only display non-duplicated items** - see screenshot)
{F542582}
- Verified in a browser that the tool-tip does not display when all answer types are valid.
{F542399}

Reviewers: amy, jared, emily

Reviewed By: amy, jared, emily

Subscribers: alex, kitt, lauren

Differential Revision: https://phabricator.khanacademy.org/D27790