Fix annoying scrolling in zooming KaTeX
Summary:
https://app.asana.com/0/132409062806753/138402849281326

To be honest, I don't fully understand all the inner happenings
causing this issue. For some reason, tall KaTeX content causes
a vertical scrollbar when we apply `overflowX: 'auto'`. Checking
in the DOM inspector shows the height of the parent and
the child is the same though, but the parent's
scrollingElement is larger by a few pixels.

Test Plan:
 - open [[ http://localhost:8080/math/precalculus/limit_topic_precalc/limits_precalc/e/finding-limits-numerically | an exercise ]] and [[ http://localhost:8080/science/physics/fluids/density-and-pressure/a/pressure-article | an article ]] with tall math (in chrome emulation iPhone 5s)
 - scroll through and make sure that the equations aren't cutoff visually
 - tap to zoom to make sure it looks fine in both zooms
 {F542950}

Reviewers: shadaj, kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27930