Fix `PropType` validation in `ItemRenderer`
Summary:
`ItemRenderer` is part of Perseus's public API and can be provided with any object for its
`apiOptions` prop. As such, we can't use the `apiOptions.propTypes` `PropType` in that component--
`ItemRenderer` itself is responsible for filling in the defaults on that object and making sure
that `apiOptions` satisfy the type in child components.

On Asana @ https://app.asana.com/0/132409062806753/143078277224746.

Test Plan:
- Open up the Renderer demo.
- Verify that there are no `PropType` warnings.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28082