Clean up unused `dynamicHintsArea` feature
Summary:
In D26381, we enabled the "dynamic" hints feature in Perseus, which allowed for hints to be placed
above the question area in some cases. This behavior was reverted in D27159 (see that diff for
rationale); this diff is a follow-up to clean up the unused code.

On Asana @ https://app.asana.com/0/132409062806753/130814462486058.

Test Plan:
- Grep for `dyanmicHintsArea` here and in webapp.
    - Verify that there are no occurrences.
- Substate this change in to webapp; do a bunch of exercises, take hints, etc.
    - Verify that hints and general exercise mechanics work as before.

Reviewers: jared, emily

Reviewed By: jared, emily

Differential Revision: https://phabricator.khanacademy.org/D28116