Update "Show another hint" text to use Proxima
Summary:
This diff updates the "Show another hint" text to use Proxima. Unfortunately, it seems like the
Proxima `+` is not perfectly centered, and so since the `+` is rendered with a different font size,
this required some manual adjustments to make the layout pixel-perfect.

The styles for the `linkButton` were taken from `phone-help-bar.jsx` in webapp (commit #914bf133).

On Asana @ https://app.asana.com/0/132409062806753/143809161382834.

Test Plan:
- Substate into webapp.
- Verify that the "Get another hint" type renders in Proxima, with the same sizing, etc., as the "Get some help" type.

Before: {F543882}

After: {F543883}

Reviewers: jared

Reviewed By: jared

Differential Revision: https://phabricator.khanacademy.org/D28236