Avoid rendering zoomed images with transparent backgrounds
Summary:
As of D28431, we render the backdrop behind zoomed-in images as an opaque gray, rather than white.
As a result, any images with transparent backgrounds end up rendering directly atop the article
content, making them very difficult to read. This diff sets the default background for zoomed
images to white, so that the image appears just as it does in the article body.

On Asana @ https://app.asana.com/0/132409062806753/146911204514724.

Test Plan:
- Substate this change into webapp.
- Open up an article ([link](http://localhost:8080/science/chemistry/atomic-structure-and-properties/introduction-to-compounds/a/polyatomic-ions)).
- Tap on the first image (chalk) to zoom (which as a non-transparent background already).
    - Verify that the image renders as before.
- Tap on the second image (molecule) to zoom (which has a transparent background).
    - Verify that the image renders with a white background.

Reviewers: david

Reviewed By: david

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28551