Apply new styling to MC items on tablets
Summary:
Unfortunately, it's a bit tricky to guard this against a flag. Ideally
the `responsiveStyling` flag applies to phones for everyone,
but only to tablets for users in `web_tablet_exercises_beta`, but
that requires a bunch of duplication, and I didn't feel it was worth
the complication. And while the new styling is cleaner and in line with
the designers' specs, I don't think it will likely affect metrics noticeably.
So this change applies the new tablet designs for all tablets.

Test Plan:
 - using [[ http://localhost:8080/math/probability/probability-and-combinatorics-topic/decisions-with-probability/e/using-probability-to-make-fair-decisions | an exercise with MC ]] ...
 - load on iPhone 5 chrome emulation:
 {F545182}
 - load on iPad chrome emulation:
 {F545181}
 - load without emulation:
 {F545183}

Reviewers: charlie, jared

Reviewed By: charlie, jared

Differential Revision: https://phabricator.khanacademy.org/D28596