Make the height of the noframe container match the height of the iframe
Summary:
Previously, the height of the noframe container was fixed, so adding too much content would introduce a scroll bar. Now, the height of the contents of the iframe is read and used to resize the parent container. This was changed because content editors need a way to be able to see all of the content at once with no scrolling, so that they can take screenshots of pages and have the screenshots contain the full content.

See: https://app.asana.com/0/144847879372674/148809046749949

Test Plan:
- Verified in a browser that adding additional content causes the noframe container to be resized (load the perseus demo, type in multiple lines of text, see that the height of the iframe container changes to match the height of the content)
- Verified that the resizing behavior is only enabled for the noframe device and is properly enabled and disabled

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28682