Disallow third-party images from being added via the Perseus content editor
Summary:
Addresses https://app.asana.com/0/148135506045999/148135506046007.

We shouldn't allow third-party images to be used in our content, because (a) those images won't work for zero-rated users, and (b) those images are out of our control, so they could be removed/changed/defaced without our knowledge or permission.

This change tweaks the Perseus content editor to disallow third-party image URLs, and adds an explanatory error message if the user attempts to use such URLs. Note that it *is* still possible to add third-party images via markdown, but addressing that is out-of-scope for this project.

Test Plan:
* Visit a Perseus content editing page, such as `http://localhost:8080/puppies/benkraft-sandbox/benkraft-subtopic/a/benkrafts-test-article/edit#`
* Change the image to a web+graphie link (e.g. `web+graphie://fastly.kastatic.org/ka-perseus-graphie/7663743cf127dc9eb1d0c039632e2bf5cc91ab51`) and ensure that the change takes effect
* Change the image to a kastatic link (e.g. `https://fastly.kastatic.org/ka-perseus-images/e988a18f4dd9efe4c51cbdee4f9a61c1dc8fe544.svg`) and ensure that the change takes effect
* Change the image to a ka.org link (e.g. `https://www.khanacademy.org/images/khan-logo-vertical-transparent.png`) and ensure that the change takes effect
* Change the image to a third-party resource (e.g. `http://cf.ltkcdn.net/cats/images/std/159706-357x421-kitten.jpg`) and ensure that the change does *not* take effect, and that an error message is shown.

Reviewers: emily, carter

Reviewed By: emily, carter

Subscribers: benkraft, carter, jared

Differential Revision: https://phabricator.khanacademy.org/D28643