Split widgets into "basic" and "extra"
Summary:
[[https://app.asana.com/0/151778147182266/151778147182267 |asana]]

This is to prepare for lazy-loading the extra widgets.
I also refactoring things so that `require("./all-widgets")` is no longer
side-effectful, which will allow me to, in the future, only load a subset of
the widgets.

Test Plan:
`make build`
`make serve`
substate in webapp, do some problems, see that it works.

Reviewers: emily, kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28772