Enable `no-undef` in all files
Summary:
With the upgrade to Babel 6 (see: D27358), all of our JS is now being run in strict mode. This has
led to some `undefined` errors -- specifically, in situations in which our JS otherwise would've
passively assigned to a global.

This diff removes all of the `no-undef` lint ignores and fixes the existing violations (or
explicitly identifies `globals` when necessary), so that we'll no longer fall victim to strict-mode's
strictness.

Test Plan:
- `make build`
- Run `make serve`.
    - Choose the "All widgets" template.
    - Verify that there are no console errors.
- Grep for `no-undef` -- verify that no files (apart from some tests in `react-components`) include that ignore.

Reviewers: emily, kevinb

Reviewed By: kevinb

Subscribers: david, jared

Differential Revision: https://phabricator.khanacademy.org/D28613