Make iframe containers render with an initial height of 100% so that content takes up the full height of device frames
Summary:
Previously, non-seamless iframes would initialize with no set height, so content would be clipped in non-seamless frames. Now, iframe containers are rendered with a height of 100% so that iframes are not clipped.

https://app.asana.com/0/144847879372674/153100318881732

Before: {F547685}

After: {F547686}

Test Plan: - Verified in webapp that content is not clipped in non-seamless frames

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28948