remove CSS rules for <IE10
Summary:
The CSS class `.ie` is applied to webapp's `body` in <IE10, and other classes of the form `\.ie[6-9]` and `\.lte[7-9]` are applied depending on the particular IE version.
Because we no longer support <IE10, we no longer care how anything on the site behaves in <IE10. We can therefore safely remove all CSS rules that target these classes, and can even remove those classes from the `body`. Hooray!

This is the Perseus change. It's small! :D There's a small corresponding webapp change, though the dependency is weak, because these two changes only affect <IE10 behavior anyway, so it doesn't even matter whether we keep these changes in sync. D28909

Test Plan:
Run `make serve` in the Perseus directory, and visit http://localhost:9000/
Create a Plotter widget, and confirm that it doesn't somehow look totally whack in a supported browser.

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28906