Enable lazy-loading for most of the widgets
Summary:
Using webpack's module splitting magic to do the heavy lifting. I fear this
will run afoul of perseus's clients though :( I'll need to mess with webapp to
see how to get it going.

With the `require.ensure`, webpack creates `build/perseus.js` *and*
`build/1.extra-widgets.js`, and when you get to the lazy-loading part, it tries to grab
that other file. However, this requires knowing what URL it will be available
at...

Now, in webapp's `javascript/perseus-package/perseus.js`, we just have a
`require()` w/ the relative path to the compiled `build/perseus.js` file from
this repo.
What would happen if we instead had webapp do all the packaging? (e.g. not
reference the compiled bundle, but instead the actual source?) I don't think
our custom packaging solution is able to do the fancy module splitting.

BUT if we pre-load the extra file (load it in a script tag on the page), then
the webpack runtime will see it & not try to jsonp stuff, and all will be
well.

So we can just
- load the main perseus js
- later (after exercise is done rendering?) add the secondary file to the
  page, so that it's loaded and ready in case there's a hint or something (or
  a subsequent problem) that needs the extra widgets.

And we'd also want to be able to disable (or highjack) webpack's default
jsonp-loading machinery.

Test Plan:
`make build`
`make serve` - using the local editor should work
and you should be able to add a dropdown or something

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28799