Semi-restore `onePerLine` property for radio widgets in editor
This restores the `onePerLine` default prop so that old content
can have it defaulted on "upgrade". It doesn't expose it in the editor
UI, though, so new content still won't be able to set it.

This was caught in the unit test, and fixes the test breakage.

Auditors: david

Test Plan:
 - `make test`
 - load [[ http://localhost:9000 | the perseus editor ]], add a radio
 option and make sure there's no one per line option