Remove Khan.scratchpad, move it into an apiOption
Summary:
Depending on this global is sad, and it's really hard to integrate
with this from inside of a redux store. So, removes calling `Khan.scratchpad.*`
and adds an apiOption function that can be passed in and called when these
things are to be used.

This will affect webapp, and shouldn't be landed until a corresponding diff is
added. It won't affect mobile because they don't depend on `Khan.scratchpad.*`
stuff.

Test Plan:
 - `make serve`
 - Visit http://localhost:9000/
 - Add an `example-graphie-widget` widget. See that the
   "Scratchpad:enabled" at the top of the screen turns to
   "Scratchpad:disabled", and there are no JS errors.
 - Repeat with `grapher`, `interaction`, `number-line`, and `simulator`.
 - Also, add `interactive-graph`, `measurer`, `plotter`, and `transformer`
   widgets and notice that there are no JS errors, but the scratchpad stays
   enabled (they do this on purpose).

Reviewers: kevinb, charlie, samiskin, jared

Reviewed By: charlie, samiskin, jared

Subscribers: oliver

Differential Revision: https://phabricator.khanacademy.org/D29242