Load extra widgets when option is undefined
Summary:
When we enabled lazy loading for Perseus widgets (D28799), we modified the API of `init` such that
we only loaded the 'extra' widgets when the `loadExtraWidgets` parameter was truthy. However, we
didn't bump the Perseus major version number. As such, our mobile clients have been continuing to
initialize Perseus by excluding the `loadExtraWidgets` parameter, even though they do need those
widgets, and thus failing to render any exercise that includes extra widgets.

This diff loosens the coercion around `loadExtraWidgets` such that we now load the extra widgets if
the parameter is not provided, making the API backwards compatible (and putting the onus on clients
to explicitly decline to load the extra widgets). We may want to instead bump the major version
number, but this will serve as a stopgap fix until then.

Test Plan:
- In the Android codebase, enable local Perseus loading.
- Substate this change into webapp (and `make serve`).
- Run the Android app.
- Open any exercise that includes extra widgets (e.g., "Performing translations" or "Making picture graphs 1").
    - Verify that the exercise loads without error.
    - Verify that the exercise renders correctly.
    - Verify that the exercise is completable.

Reviewers: jared, benkomalo, william

Reviewed By: william

Subscribers: #ios, #android

Differential Revision: https://phabricator.khanacademy.org/D29269