Added add and remove highlight confirmation icons to SAT Highlighting
Summary: Now, when a user wants to highlight a new section they have to confirm by clicking on an add highlight icon. In addition there is also now functionality to remove the highlight by clicking on the highlighted section and then on the remove highlight icon.

Test Plan:
Verify the ability to add and remove highlights from a page using the icons.

{F549225}
{F549226}

Reviewers: sean, mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D29442