restrict XOM phone styles to XOM
Summary:
Because BaseRadio is also used in SAT, a style change in D29147 caused a visual regression in SAT radio buttons: applying `overflow: auto` to the container caused `outline`s to no longer visibly overflow beyond the vertical boundaries of the radio widget.
This change gates those styles behind the `responsiveStyling` Perseus API option, which is `false` in the SAT product, but, I think (can someone from XOM confirm?), is `true` everywhere that X-On-Mobile styles are applied.

Test Plan:
In an SAT multiple choice item, choose "A". Confirm that the top edge of the blue outline is visible. {F549660}
In an SAT multiple choice item. choose "D". Confirm that the bottom edge of the blue outline is visible. {F549664}
In a content library exercise on a small screen, use DevTools to make one of the answer choices suuuuuper long. Confirm that the radio widget provides a horizontal scrollbar. {F549661}

Reviewers: david, benkomalo

Reviewed By: benkomalo

Subscribers: riley, amy, jared, #sat

Differential Revision: https://phabricator.khanacademy.org/D29514