Enable custom keypad for number line
Summary:
The number line widget has a single input, which is used to control the number of divisions in some
cases. This diff updates that widget to use the custom keypad, when specified.

There's one oddity, which is that our input is a controlled component, and so when the user deletes
all the text in the input, we need to update the `value` prop that's passed to our component to
reflect the empty text. I had to track this state internally, since before, the number line
wouldn't update at all when the input became empty.

On Asana @ https://app.asana.com/0/132409062806753/157274192695176.

Test Plan:
- Open up a number line ([link](http://localhost:9000/#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20number-line%201%5D%5D%5Cn%5Cn%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22number-line%201%22%3A%7B%22type%22%3A%22number-line%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22static%22%3Afalse%2C%22range%22%3A%5B0%2C10%5D%2C%22labelRange%22%3A%5Bnull%2Cnull%5D%2C%22labelStyle%22%3A%22decimal%22%2C%22labelTicks%22%3Atrue%2C%22divisionRange%22%3A%5B1%2C12%5D%2C%22numDivisions%22%3A5%2C%22snapDivisions%22%3A2%2C%22tickStep%22%3Anull%2C%22correctRel%22%3A%22eq%22%2C%22correctX%22%3Anull%2C%22initialX%22%3Anull%2C%22isTickCtrl%22%3Atrue%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22calculator%22%3Afalse%2C%22chi2Table%22%3Afalse%2C%22periodicTable%22%3Afalse%2C%22tTable%22%3Afalse%2C%22zTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D)).
- Open up the rendered view.
- Tap to reveal the custom keypad and the cursor.
- Tap to delete the single number in the input.
    - Verify that the cursor disappears, as does the number.
- Verify that the desktop behavior is as-before.

{F549769}

Reviewers: jared, kevinb

Reviewed By: jared, kevinb

Differential Revision: https://phabricator.khanacademy.org/D29516