Fix a bug where invalid inputs are considered empty
Summary:
https://app.asana.com/0/132409062806753/161594397687316

This is the first step in fixing the exercise pages from disabling
the `Check` button. The next step is to do a bit more in the cases
where an attempt message is actually sent back up the exercise.

Initially I tried this by refactoring the `score` object from
`src/util/answer-types.js`, but realized that was a losing battle;
the "KE score" object is converted to a "Perseus score" object
in all of the individual widget implementations, and those score
objects too have an aliasing problem where multiple states
are hard to distinguish. One day I'd love to move away from these
untyped object literals (especially in the absence of `Flow`), but
right now it's a bit hard/scary to refactor those. I opted instead
to make the assumption that a presence of an attempt message implied
that the widget is not strictly empty, but instead had some other
issue.

Note this changes the behaviour of a publicly-facing method on
Perseus, though arguably it's a bug fix and not a behaviour change,
so no API version bump is made.

Test Plan:
 - in the host webapp, load [[ http://localhost:8081/math/pre-algebra/fractions-pre-alg/mult-fracs-word-probs-pre-alg/e/multiplying-fractions-by-fractions-word-problems | an exercise with fractions ]]
 - on both desktop and phone (new exercise framework)
 - get a question wrong
 - get it right after getting it wrong
 - get it right on the first try
 - answer a problem such that you have a fraction not in lowest terms

note: this causes a bug on phones on XOM+new exercise framework, because the answer can now get submitted when they have an `attemptMessage` and not graded and we didn't properly handle that case:
{F549991}

I'm going to make an immediate followup change to this, but unless y'all feel strongly, am still not going to do a version bump

Reviewers: emily, david, charlie

Reviewed By: charlie

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D29622