Update scaleChildToFit so that is measures correctly
Summary:
We were measuring once correctly on the first call to scaleChildToFit, but
some content loads asynchronously, e.g. graphics in tables so we need to
call it multiple times.  Since we’ve already scaled the table when we call it
a second time we can’t use getBoundingClientRect() which returns the scaled
dimensions.  Instead we use offsetWidth and offsetHeight which are unaffected
by CSS transforms.

https://app.asana.com/0/132409062806753/160685902998288

{F550139}

Test Plan:
- load http://localhost:8080/math/cc-fifth-grade-math/cc-5th-place-value-decimals-top/cc-5th-expanded-form-decimals/e/writing-and-interpreting-decimals
- check that math still scales
- load http://localhost:8080/science/physics/one-dimensional-motion/displacement-velocity-time/a/what-is-physics
- check that images still scale
- load http://localhost:8080/math/early-math/cc-early-math-counting-topic/cc-early-math-counting/e/counting-objects?1
- check that the table zooms correctly and does not scroll vertically

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: shadaj, jared

Differential Revision: https://phabricator.khanacademy.org/D29661