Check for unescaped dollar signs in content
Summary:
We tell content creators to escape their dollar signs in paragraphs, but
sometimes they forget.  And since it renders fine, it's hard for them to
know they've even done a thing wrong.  But there's a DARK SIDE.
Behavior that lurks in the shadows!

The translation linting will FAIL on the translated string because it
will think that the unescaped $ is a BROKEN mathjax expression!!!!

Such content will be left untranslatable, causing grief, tears, and
sadness.

In this diff, we find unescaped dollar signs in perseus content, and
generate a warning on save that will hopefully not be ignored.

All the thanks to Oliver, who was my co-pair-programmer on almost all of
this. <3

Test Plan:
    - make test
    - Try editing an exercise to have some unescaped $s, and see the
      warning

Reviewers: emily

Reviewed By: emily

Subscribers: oliver, jared

Differential Revision: https://phabricator.khanacademy.org/D29517