Implement default copy/paste
When handling text being pasted in, we must parse it into proper
Draft.js ContentBlocks.  This diff implements what Draft.js' default
paste behavior is (as if we "return true"'d it').  The reason this had
to be reimplemented was because for the custom behavior, we'll need to
be able to change certain parts of this default behavior