Implemented dropping of text and links
Summary:
This handles the situation of, in another browser, dragging in an image.
This inserts the appropriate image link for that image

The images are placed in a new block after the block that the cursor is
currently on.

Test Plan:
Open blank perseus, start typing some of a line, then place cursor in the center
Drag in image
Observe proper `![](....jpg)` insertion after the current line
Drag in text
Observe proper text insertion

Reviewers: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D29938