Disable responsive styling for SAT
Summary:
D29861 resmoved the `responsiveStyling` flag, but some of the styling
in there conflicts with SAT theming. This is a bit unfortunate, so
the easiest solution is to just turn it off for SAT.

Long term, it feels like this should be renamed internally, because
being "responsive" is just a generic trait that seems universally
desireably, so there are things associated with the "responsive styles"
that are more to do with theming that should be separated.

Test Plan:
 - load any SAT math practice task on a window that isn't `xlarge`
 {F551319}
 - load any normal exercise with MC
 desktop: {F551321}
 tablet (XOM): {F551320}

Reviewers: kevinb, mattdunnrankin

Reviewed By: kevinb, mattdunnrankin

Subscribers: emily, jared

Differential Revision: https://phabricator.khanacademy.org/D29980