Webapp fixes
Summary:
Certain issues happened when running perseus in webapp for cases
such as clicking a different assessment item, or parsing an assessment
item which had extra widgets.  This fixes those.

Test Plan:
Open an exercise in `devadmin/content`, click some of the assessment
items and observe the editor opening and updating correctly, rather than
either erroring or not updating the text for different assessment items

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D30272