Made the selections snap to outside entities
Widgets should be considered to be equivalent to a single character.
This means you should not be able to have your cursor inside one, and
you shouldn't be able to delete only part of a widget

To allow this, I set it to always snap the selectionstate to surround
any entities that are at the edges

Test Plan:

Add some text with widgets in it

Make a selection starting at partially in one widget, and ending in
partially in another widget.  Observe the selection filling the two.

Make a selection in the middle of a widget, observe the selection
expanding to fill the entire widget

Place the cursor inside the widget, observe it snapping to the rigth
Press Left, observe the cursor skipping to the left end of the widget
Press right, observe the cursor skipping to the right end of the widget