Fix preview frame in firefox
Summary:
I opened a perseus editor in firefox and the frames didn't seem
to be resizing appropriately. I saw there was a JS error, and found that
a call to `document.querySelectorAll().forEach` (which doesn't always
exist, because it is a NodeList not an Array) wasn't working. This calls
`Array.from()` to use the `.forEach` function.

Test Plan:
 - `make build/frame-perseus.js`
 - Visit
   http://localhost:8081/math/algebra2/exponential-and-logarithmic-functions/graphs-of-exponential-functions/e/graphs-of-exponential-functions/edit/x45f029ec7ee8cf21
   in firefox
 - See that the frames resize reasonably.

Reviewers: benkomalo, kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D30122