Render device previews w/o device chrome + use proper device sizes
Summary:
The device chrome uses up a lot of extra space in the browser which
content creators may not have if they're working on a laptop.  This
diff renders previews that fit within a laptop's screen area on a
MBP where Chrome takes up the full screen.

https://app.asana.com/0/132409062806753/169651639018834

{F551706}

{F551707}

{F551708}

Should be deployed with https://phabricator.khanacademy.org/D30127.

Test Plan:
- load http://localhost:8080/quetzalcoatl/cameroon/middle-school-content/e/ideas-rhonda/edit/x86c3dba4b17f5fc5/preview
- switch between the different device viewports: phone, tablet, desktop
- check that switching questions still works

Reviewers: jared

Reviewed By: jared

Differential Revision: https://phabricator.khanacademy.org/D30126