Fixed selection around entities bug
Summary:
If the user had their cursor at the right side of an entity, then
pressed the left arrow, the cursor would immediately snap to the right
side.  To deal with this I refactored snapSelection to use focus/anchor,
since if the user used the arrow keys it would always move the focus (if
they were in a selection and pressed shift-left, it would move the focus
but not the anchor).  If the focus moved by one, snap it to the opposite
side based on the direction it moved

Test Plan:
Create basic set up with some widgets
Place the cursor to the right of a widget, press left and watch it skip
over the widget
Repeat above from the left side

Place to the right, hold shift, and start pressing left, and watch the
selection jump to the left side of the widget.  Still holdign shift,
press right again and watch it jump back over the widget
Repeat above from the left side

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D30276