Fix touch emulation to work with Movable elements
Summary:
Movable's update method does a sanity check to see if 'which' is either 0 or 1
before proceeding.  The touch emulation code wasn't setting the 'which' property
so this check failed.  This diff adds the 'which' property to the simulated touch
events which makes the interactive elements work in the editor.

https://app.asana.com/0/132409062806753/169651639018843

Test Plan:
- open an exercise in the editor e.g. /e/addition_1/edit
- add an interactive graph, number line, and plotter
- interact with them in the edit tab by dragging the points around
- switch to the preview tab and verify that dragging works there too

Reviewers: jared, benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D30161