Made the editor undecorate if the selection is already decorated
Summary:
If the entire selected region is already decorated, pressing the
hotkey should undo the decoration.  To deal with this I had to check every
block for whether it should be decorated, but I didn't want to have to
duplicate the code very much.  Since this task only gets ran once when the user
clicks ctrl-b, I figured it was worth the performance impact to just do both
tasks of decorating and undecorating, then if all the blocks were supposed to
be undecorated, using the undecorated version

Test Plan:
PAste in lorem ipsum with 5+ paragarphs.  Select from the middle of one to the
middle of another, with muliple paragraphs in between.  Press Ctrl B, observe
all of them bold correctly

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D30282