Disable image zooming on exercises
Summary:
In practice, this isn't useful. It's pretty rare that a piece
of exercise content has a large enough image that it needs to be zoomable,
so the default behaviour is rather awkward.

Moreover, it doesn't play well with the way the native app renders
the layers of the widgets. :( So disabling it will help avoid oddities
with a full-screen zoomable interacting badly with that.

https://app.asana.com/0/143216769809283/186311201300445

Test Plan:
 - temporarily make an Android build to point to localhost (see `perseus-common.js` in its webview repo)
 - load the exercise "Making 5" in Early Math > Addition and subtraction intro
 - tap on the image, make sure it doesn't zoom

Reviewers: david, jared

Reviewed By: david, jared

Subscribers: cameron

Differential Revision: https://phabricator.khanacademy.org/D31068