Passage widget: only measure line height once per mount/resize
Summary:
https://app.asana.com/0/110018685008094/193494917573040

Some users have noticed poor performance in IE and Firefox when working on reading passages.
From my experimentation in Firefox's profiling tools, it looks like the many calls to `measureLineHeight` were a source of slowdown - about 3 seconds lag on my Macbook, so presumably an older machine could take significantly longer and time out, which would explain the tickets we're seeing.

This change caches the line height measurement on `LineHeightMeasurer`, and has the `Passage` widget explicitly signal that `lineHeightMaybeChanged` on resize events, because zoom events are currently the only time in the `Passage` widget's lifecycle that line height will change.
Now, we only call `measureLineHeight` once on mount, once on window resize (and the resize handler is throttled), and zero times on other `Passage` updates - whereas, before, we were calling it like a bajillion times (maybe 20-30 from eyeballing my console) per Perseus update. There's still //some// lag when I click a radio button in a reading passage in Firefox, but it's nowhere near the 1-second delay that it was before, and the profiler doesn't seem to be pointing to any particular line of code as the culprit anymore, so I vote that we ship this and see if we at least stop timing out for those users who reported slowness.

Test Plan:
In Firefox, open a reading passage task.
Select one of the radio buttons.
Observe that the window unfreezes and allows you to scroll within a second, instead of the 3 seconds that it currently takes on prod.
For good measure, add a `console.log` line to `measureLineHeight`, and confirm that it's only called once per mount, once per resize, and zero times per other update.

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D31283