Clear out widget IDs during unmounting
Summary:
We're getting a bunch of `null` errors from this line in `renderer.jsx#getInputPaths()`:

```
_.each(this.widgetIds, (widgetId) => {
    var widget = this.getWidgetInstance(widgetId);
    if (widget.getInputPaths) {
        ...
```

The issue is that `this.getWidgetInstance(widgetId)` is returning `null`, which implies that the
renderer is in an inconsistent state (i.e., the available refs don't match the values stored in
`this.widgetIds`). I have no evidence that this is the right fix, since I can't repro this issue,
but it seems _possible_ that we're calling `getInputPaths` after the renderer has been torn down
(maybe on slow devices, or something, where the native<->JS communication is delayed?). In any
event, it seems like the renderer is left in an inconsistent state after unmounting, and so this
should lead to saner behavior?

Very open to suggestions here :)

Test Plan:
- Open up an exercise in the editor.
- Tap around on it; make sure nothing crazy happens.

Reviewers: emily, alex, jared

Reviewed By: alex, jared

Subscribers: #ios, #android

Differential Revision: https://phabricator.khanacademy.org/D32420