[multirenderer] hints!
Summary:
https://app.asana.com/0/211627504842249/225424135385019

In this change, we add a new primitive leaf node, `shapes.hint`.
In the editor, we use a simple `HintEditor`.
When rendering, the layout function can either render a single hint directly via e.g. `renderers.hint`, or, if there's an array of hints, the layout can render the first N of them in a single `HintsRenderer` by calling `renderers.hints.firstN(n)`.

Test Plan:
I did edit, preview, and JSON, and it's all happy :) I'm about to leave the office so this is a lame description, but I did it :P
{F595378}

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D32580