[multirenderer] quick sanity improvement on pluralToSingular
As recommended in code review, but I forgot to do it :P
We now check that the string ends with `s`. If not, we just return the plural form.

Test Plan:
http://localhost:9000/?multirenderer
Singular "Question" and "Hint" look good.
When I edit the shape key to be `questionsYay`, the first question header is "Questions yay 1", as expected.

Auditors: emily