[multirenderer] Modify all of our tools to not break on multi-items
Summary:
We have a bunch of i18n, a11y, etc. tools that we use from
webapp to modify or inspect perseus assessment items. Using the `_multi`
key that is at the root of all multi-items, this makes all of the tools
return sane defaults on multi-items, and adds tests for those things.

Depends on D32740

Test Plan: - `make test`

Reviewers: mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: jared, #sat

Differential Revision: https://phabricator.khanacademy.org/D32741