Guard against re-tearing down Zoomable in callback
Summary:
According to Sentry, `this._activeZoom` is occasionally `null` here, which seems rare but possible
given that we're tearing it down async, and it could've been torn down before the callback was
triggered (e.g., if the user tapped outside of the image before the callback fired, we'd immediately
dispose of the Zoomable and then error here).

On Sentry @ https://sentry.io/khanacademyorg/android-webview/issues/200552426/.

Test Plan:
Cross fingers? It would be high-effort to test this and the change is safe / defensive,
but happy to go through the motions if there's a concern.

Reviewers: david, benkomalo

Reviewed By: benkomalo

Subscribers: benkomalo, #android, jared, #ios

Differential Revision: https://phabricator.khanacademy.org/D32874