[multirenderer] Add question-level tag shapes.
Summary:
This adds the ability for shapes to indicate that they'd like to
include assessment item tags. In SAT, we use these tags to distinguish between
the different questions in a passage item with multiple questions, but others
might also find this helpful.

I also noticed that there's a bunch of disjoint places that we deal with adding
a new type. There might be a better way to do this, if we imagine that we're
going to be adding more types in the future...

Test Plan:
 - `make test`
 - Visit http://localhost:9000/?multirenderer
 - Add a question
 - See that a tags field appears and that I can add elements to it and delete
   elements from it, and there are no JS errors.
 - Switch to the JSON mode, see that the tags appear in the JSON.
 - Switch to the preview mode, see that there aren't JS errors.

Reviewers: mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: lauren, jared, #sat

Differential Revision: https://phabricator.khanacademy.org/D33030