[multirenderer] refactor into a legit library with modules, docs, and types
Summary:
And awaaay we go! :D The goal of this refactor is to bring our abstract concepts like items, shapes, and item-like trees into the spotlight with clear modularity, typing, and documentation, so that future readers will be able to dive into the piece they care about with minimal overhead.

For now, this is a strictly internal refactor. The external API is left untouched (with some thin glue code to make that happen), and the unit tests still target that external API. Later this week, we should target updating the external call sites and moving the tests inside :)

We also don't yet have support for the tag primitive. That'll probably be the next target once this draft is done. The highest surface-area effect of adding tags will be that the `Tree<C, H>` type will receive a third type parameter, which won't actually complicate any *logic*, but might be a bit more verbose than we'd like - but that's sorta just the cost of doing business when you're type-checking generics, ya know?

Okay, cool. For the rest, I'm just gonna let the docs speak for themselves. Please pass along copy edits as well as code edits - these docs are tricky and getting them right is important!

Test Plan:
`make test` (which includes Flow checking :D)
I also tested the core features of the Perseus demo editor.
Haven't tried webapp integration yet, though, but I'm pretty confident we're okay since the unit tests pass, and we'll end up doing plenty of manual testing here when we update the call sites to use the new API.

Reviewers: emily, kevinb

Subscribers: lauren, jared

Differential Revision: https://phabricator.khanacademy.org/D33138