Add label "Answer" next to numeric input widget in SAT tasks
Summary:
Kitt would like there to be a label next to input fields in SAT practice. I'm restricting to SAT using satStyles flag.

Answering a fill in the blank question:
{F607908}

After submitting:
{F607912}

(Note: I left the after submission as is since I thought it looked better than having the double labels. but LMK if you disagree @kitt)

Test Plan: confirm we do see the "Answer:" label in SAT but not in other text input fields on other parts of the site (e.g. an algebra mission)

Reviewers: emily, mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: kitt, lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D33286