Prevent highlighting logic from indexing out of bounds in passage text
Summary:
Fixes https://app.asana.com/0/110018685008094/275915325056246

Within the SAT product, we've seen a stream of recent reports about answers going missing and passage navigation breaking. I was able to reproduce the problem by making highlights on the very last line of a reading passage. By doing so, I was able to get into a state in which the index we've recorded for the end of a highlight range exceeds the length of the passage's text array (leading to an uncaught javascript exception in perseus, and behavior similar to what users have reported). I have not uncovered the root cause of the last-line bug, but it is presumably an edge case is in the (very hacky) way we handle highlighting ranges (based on timing of the regression I think it is related to our recent React upgrade, which we know affected highlighting logic in other ways as well). Ultimately we need a full refactor of the way we do highlighting in passage widgets. But for now, I just ensure that we don't try to access out of bounds of the text array since that completely breaks the passage widget experience.

Test Plan:
confirm that without the change, I can repro reported behavior of highlights not showing up, selected options not showing up, and navigation between passages being broken
confirm that with the change, the same attempt data renders correctly (answers and highlights are visible) and uncaught exception is gone
confirm that highlight continues to work as expected (expected ranges are highlighted)

Reviewers: emily, mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: sergei, lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D33799