Add some additional documentation around version bumps with respect to mobile
Summary:
There are a bunch of undocumented, implicit dependencies between the mobile clients and Perseus,
which require us to bump the Perseus major version number in certain situations so-as to avoid
mobile breakages. This is an attempt to document some of those situations. Note that this list is
probably not exhaustive, but I think it's fairly comprehensive.

Test Plan: N/A

Reviewers: emily, amy, mattdunnrankin, jared

Reviewed By: emily, amy, mattdunnrankin, jared

Differential Revision: https://phabricator.khanacademy.org/D34015