Avoid rendering static overlay above exercise footer
Summary:
Currently, the static overlay sits above the exercise feedback popover (rendered by webapp),
stealing clicks on the popover whenever it's active. The popover sits at a z-index of
`interactiveComponent`, so I've just brought down the z-index of the overlay to
`interactiveComponent` as well to ensure that it'll render above any interactive elements in
the widgets, but below the popover. (The z-index situation between Perseus and the rest of the site
is a bit fraught. Perhaps the z-index of the footer + exercise popover should just be increased to,
say, 4 (in addition to this change), but I'm worried about unintended effects.)

On Asana @ https://app.asana.com/0/16941950784229/293427273100249.

Test Plan:
- Substate this change into webapp; run `make watch` here and `make serve` in webapp.
- Open up http://localhost:8080/science/physics/one-dimensional-motion/displacement-velocity-time/e/average-velocity-and-average-speed-exercise.
- Get a question right.
- Scroll such that the feedback popover overlaps the radio widget.
    - Verify that you now tap the 'X' button to close the modal.
- Edit the DOM to enable the radio widget choices.
    - Verify that you _cannot_ click on them (i.e., that the static overlay is working as intended).

Reviewers: diedra, kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D34425