[multirenderer] Call the onSerializedStateUpdated callback before the component updates
Summary: In the previous diff, (D34463) we added a callback to `<Renderer>` to
make it tell us when serialized state changed, and to make a kinda-controlled
serializedState prop. Unfortunately, we weren't calling the callback quickly
enough, so the renderer would revert back to its pre-changed state due to the
kinda-controlled prop. This calls the callback earlier so that the parent
component actually has time to update the prop, and handles this in the
multirenderer.

Test Plan:
 - `make build`
 - See that I can change answers in the question UI

Auditors: mattdunnrankin