Making it easier to expand a draggable/resizable Graphie circle when its radius is 1.
Summary:
Asana issue: https://app.asana.com/0/253524899192327/152764110371615 -
the investigation of this issue uncovered several problems, all of which
are fixed here. Let me know if you want me to split up the reviews.

In "draw a circle" exercises, the starts out with radius 1. Because of
the size of the invisible circle that receives the click events for
moving the circle, it is very difficult to hit the circumference in
order to expand the circle. The solution is to simply reduce the click
target's size a little bit.

We are also fixing another usability problem: the div and svg that
surround the click target are square, so if you click in one of those
corners, nothing will happen (while you would expect to hit the click
target for expanding the circle). Solution: disable mouse events on
those elements.

Also, it was impossible to shrink the circle back to radius 1 after
initially expanding it.

This review also contains a suggested improvement (which I'll be happy
to revert or move to a separate review if requested): I think it would
be good to have separate cursors for resize vs. move, in order to make
it more obvious which action you'll trigger when you click. I went for
the Northeast/Southwest arrow, but it can be discussed (and maybe we
could even switch between the eight resize arrows based on the cursor
position).

Test Plan:
- `make serve`
- Visit
  `http://localhost:8081/math/algebra2/intro-to-conics-alg2/standard-equation-circle-alg2/e/graphing_circles`
-- Check that even when clicking slightly inside the circumference,
you'll be able to expand the circle
-- Check that the circle can be shrunk back to radius 1
-- Check that when the circle has radius 1, no matter where you are
inside the circle, the pointer is always in "expand" or "move" mode;
there should be no "dead" areas (in particular, check the "corners")
-- Check that the "expand" cursor is different from the "move" cursor

Reviewers: emily

Reviewed By: emily

Subscribers: sergei, jared

Differential Revision: https://phabricator.khanacademy.org/D34486