Allow for previewing rationales
Summary:
It would be helpful for content creators to be able to preview their rationales in the CMS (especially as they start entering in more complex content in the rationales). This diff handles the perseus side of wiring up the showRationales and reviewMode flags needed to actually render the rationale.

NOTE: further perseus work will be needed for the rationales to render with appropriate styles/positioning in the standard exercise chrome.

Depends on D34937 (webapp changes)

Test Plan: in coordination with the webapp changes, load up an item in the CMS and see that toggling the "enable rationale preview" toggle changes whether the rationale renders when a question has been checked

Reviewers: mattdunnrankin, emily, michaelpolyak, johnsullivan

Reviewed By: mattdunnrankin, johnsullivan

Subscribers: lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D34939