Replace throbber.gif with spinner.gif
Summary:
We replaced our loading animation on KA and want to update perseus too! No more weird mission-like box-y thing.

NOTE: there are various references to the throbber in khan-site.css but apparently that css was essentially autogenerated ages ago and not updated, so we're gonna leave that as is ¯\_(ツ)_/¯

Test Plan: Plan to test on a ZND along with the rest of the spinner changes. Will try to find a slow-loading graphie and confirm we get the new spinner

Reviewers: emily

Subscribers: lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D35038