[highlighting] First draft of new highlighting architecture
Summary:
In this change, we disable the existing SAT highlighting feature in passage widgets, and replace it with a more generalizable system that directly interfaces with the DOM, instead of mucking around in Markdown source, which is a major source of bugs and complexity in the existing system.

The key components are:

* `HighlightableContent`, which fully encapsulates the process of highlighting the DOM inside it. It takes serialized highlights as input, and yields serialized highlights to a callback on user interaction.
* `WordIndexer`, which traverses the content's DOM to find words. It yields the list of words to `HighlightableContent`, which caches it in order to quickly map serialized highlights into DOM `Range`s (for rendering output) and vice-versa (for user input).
* `HighlightsRenderer`, which applies highlights to the content by measuring the highlighted content's position and placing absolute-positioned semitransparent highlight rectangles above the content.

This module isn't even Perseus-specific—it's just DOM! It might be interesting to pull this into its own package, once it's more feature-complete :O

I left some high-level documentation TODOs lying around, because it's late and I'm tired, but I tried to do good low-level comments along the way, and hopefully this brief architecture summary is enough to get you started :) Any questions, hit me up!

[[ https://app.asana.com/0/277557989281705/318727931059822 | Asana task ]]

NOTE: Do not land until we have feature-parity with the existing SAT highlighting feature.

Test Plan:
http://localhost:8081/prep-for/lsat/task/TASK_ID/question/1
Use Shift-Click to highlight some areas of the passage.
(Dragging also works, but will trigger lots of select events and create lots of highlights, sooo maybe don't do that yet :P)
Click a highlight to delete it, too.
Wait a few seconds for task state to persist to the server.
Refresh, and confirm that your highlights are re-applied.

{F683990}

Reviewers: amy, emily, nickbreen, joshua

Reviewed By: amy, nickbreen, joshua

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35100