Make perseus-widget-passage flexible width
Summary: Currently the perseus passage widget is a fixed with of 420px. This works in the SAT product which has a fixed viewport size and isn't responsive, but doesn't work for the new test prep layout. Luckily, the fixed width doesn't seem to be necessary even for the SAT styles, so we can get away with supporting the flexble width layout without any special-casing of SAT.

Test Plan:
- confirm that passage now renders at a flexible width within the responsive layouts in test prep product
- confirm that SAT layout still looks reasonable across screen sizes (I couldn't notice a difference)
(I think SAT is only current consumer of passage widget, pls lmk if you know of others!)

Reviewers: mattdunnrankin, emily

Subscribers: lauren, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D35277