[highlighting] Fix highlight displacement bug
Summary:
When highlights are first rendered, the relevant Aphrodite styles aren't written to the document's CSS until a tick later. For this reason, if multiple highlights were rendered in that initial batch, they wouldn't have `position: absolute` applied yet, and would be statically-positioned in the document. This displaced the content, and caused each subsequent highlight to measure the newly-displaced content, instead of the content's correct position in the document.

This change moves `position: absolute` to the `style` attribute of the highlight rectangles, so that they're always absolute-positioned, even before Aphrodite styles land. That way, they can never displace content, and subsequent measurements will be accurate.

Test Plan:
Save some highlights to a test prep task's state, refresh the page, and watch the highlights be restored.

Before this change: {F690889}
After this change: {F690888}

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: joshua, emily, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35360