Make passage-ref widgets work inside hints for structured items
Summary:
Perseus allows content creators to add passage-ref widgets to hints
but when the hints are rendered, the passage-refs just come out as '?-?'.
This patch fixes that. The problem was that the hint-renderer component
was not getting passed a findExternalWidgets() function that it could
use to look up the passage widget to which the passage-ref refers.

Test Plan:
- Edit an LSAT question to add a passageref to hints
- Verify that the hint displays in the editor preview, preview and on site

Reviewers: emily, mattdunnrankin

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D35724