[cp mcr] Don't show correct answers for numeric inputs in non-sat review mode
Summary: In the SAT product, since students don't have to get the answer
correct in order to progress to the next question, we show them what the
correct answer would have been for a numeric input question.

This feature was currently gated only behind `reviewMode` being turned on, but
we only want to show it in the SAT product. So, this diff gates showing the
correct answer behind both `reviewMode` and the `satStyling` flag.

In the future, we might want to show something like this to more people, since
this does seem useful. However, we don't have designs for it and don't want for
it to be weird, so I'm just turning this off for now.

Test Plan:
 - `make watch`
 - In webapp, visit http://localhost:8081/mission/sat/practice/math
 - Open a "Solving quadratic equations" practice task
 - Get the answer wrong, and see the "CORRECT ANSWER:" prompt show up.

Auditors: amy, mattdunnrankin