[highlighting] Remove key from highlight objects
Summary:
This change cleans up one of the warts of D35623, in which we had to create a temporary key for the `DOMHighlight` before it actually got added to the set, at which point we created another key.

Instead, this change moves the `key` property off of the highlights themselves, and instead starts working with `HighlightSet` objects, which map from keys to highlights. Now, the key is a property of the set, not the highlight itself.

Depends on D35623.

NOTE: This changes the highlights data format from an array to a map-like object, invalidating old saved highlights. Thankfully, highlighting v2 has shipped to no users outside of KA, so we're free to change the data format as much as we like!

Test Plan: Make some highlights, and confirm that they're still created with the expected unique keys, and stored in the new expected data format.

Reviewers: amy, nickbreen

Reviewed By: nickbreen

Subscribers: lauren, jared, joshua, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35624