[highlighting] Only highlight text, not paragraph blocks
Summary:
`Range.prototype.getClientRects` returns a set of rectangles covering //all// nodes fully contained by the `Range`—so, when you highlight a full paragraph, we draw one big highlight rectangle covering the whole paragraph, instead of highlighting all of the paragraph's text.

This change introduces a custom DOM traversal function that finds the `Range`'s text nodes and calls `getClientRects` on the ranges of //just text//, instead of on all the contained elements.

This also opens the door to some future (but lower-priority) UI improvements, like adjusting the rectangles to be the full line height, and cropping the rectangles according to the parent's `overflow` style (and maybe removing our hack from `.perseus-sr-only` :P).

Asana: https://app.asana.com/0/329800276300868/322842086114506

Depends on D35680 for the `rangesOverlap` function.

Test Plan:
Select the end of paragraph 1, the entirety of paragraph 2, and the start of paragraph 3. Save it as a highlight. Confirm that it highlights just the text of the paragraph, rather than the full paragraph block. (I also saved some additional highlights, to confirm no regressions.)

Before: {F705064}
After: {F705063}

Reviewers: amy, nickbreen

Reviewed By: nickbreen

Subscribers: joshua, lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D35688