[highlighting v2] Disallow highlights in other highlights
Summary:
Previously, if you selected text that was fully contained by an existing highlight, you would see an "Add highlight" tooltip—but clicking it did seemingly nothing, because the new highlight just gets merged into the existing highlight, and is a no-op.

In this change, we add to our highlight validation logic by checking for whether the new range is fully included by any existing highlight ranges. Now, "Add highlight" only appears when selecting text that is not already fully highlighted.

Asana: https://app.asana.com/0/329800276300868/338136950125382

Test Plan:
Select text that is not included in a highlight, and see "Add highlight".
Select text that starts in a highlight, but doesn't end in a highlight, and see "Add highlight".
Select text that ends in a highlight, but doesn't start in a highlight, and see "Add highlight".
Select text that starts in one highlight, and ends in another highlight, but has a gap in the middle that isn't highlighted, and see "Add highlight".

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D36224