[highlighting v1 cleanup] Remove now-unused Highlighting V1 code
Summary:
When we initially launched Highlighting V2, we hid it behind a Perseus flag, which was always activated for the new Test Prep product, but only activated in the SAT product when a Gandalf bridge was crossed.

We've now gone a week with all users crossing the Highlighting V2 bridge, and no reports of problems, so it's time to start cleaning up the code!

In this change, we actually remove Highlighting V1's dead code, hooray! Turns out it was pretty straightforward :) We removed the logic, the references to it (like renders and mount hooks), and the corresponding props and state.

Incidentally, by removing all this dead code, we moved the file back into the acceptable line number limit! I therefore removed the corresponding linter exceptions, and also narrowed the scope of our `react/sort-comp` exception, so that the file no longer has any global linter exceptions. Hooray!

Depends on D36481, which updated Perseus to always use Highlighting V2. Does not depend on D36485 or D36487, which cleaned up webapp's call sites.

Asana task: https://app.asana.com/0/277557989281705/318877243057038

Test Plan: On the SAT task page, do some highlighting. Confirm that it all still works as expected, and use React DevTools to confirm that the now-unused Passage props and state are no longer present at runtime.

Reviewers: amy, nickbreen

Reviewed By: amy

Subscribers: lauren, jared, #gtp

Differential Revision: https://phabricator.khanacademy.org/D36490