[cp mcr] Implement the "disable currently selected choices" and "deselect incorrect choices that are already selected" interaction
Summary:
As part of the MCR styles, we want two interactions involving modifying the
choices as the user interacts with them. In the first, once
a user has checked their answer and we
want to disable the choices that they've already selected, since we've now told
them which of the choices are correct and incorrect. In the second, once a user
closes the answer feedback popover, we want to deselect the incorrect choices
that the user has selected, since we've now highlighted the answers they chose
for them, and we've disabled the choice and we don't want to leave the incorrect
choices selected.

These are implemented by changing the `showRationalesForCurrentlySelectedChoices`
function and adding a new `deselectIncorrectSelectedChoices()` function.
The new function is separate because we want to call
it at a different time: `showRationalesForCurrentlySelectedChoices()` is called
immediately after you check your answer, and
`deselectIncorrectSelectedChoices()` is called later after the user dismisses
the feedback popover.

Since there is no feedback popover in the demo page, we just call the two
functions one after another when you click the "check answer" button.

Gif of this in action:
{F735335}

Test Plan:
 - Visit [this item with a multi-select radio widget](http://localhost:9000/?renderer#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20radio%201%5D%5D%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22radio%201%22%3A%7B%22type%22%3A%22radio%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22choices%22%3A%5B%7B%22content%22%3A%22correct%20choice%21%22%2C%22clue%22%3A%22yay%20you%20did%20it%22%2C%22correct%22%3Atrue%7D%2C%7B%22content%22%3A%22incorrect%20choice%20%3A%28%22%2C%22clue%22%3A%22pls%20pick%20the%20other%20one%22%7D%2C%7B%22isNoneOfTheAbove%22%3Afalse%2C%22content%22%3A%22incorrect%20choice%22%2C%22clue%22%3A%22nooo%20don%27t%20pick%20me%22%7D%2C%7B%22isNoneOfTheAbove%22%3Afalse%2C%22content%22%3A%22I%27m%20also%20correct%21%22%2C%22clue%22%3A%22yay%20this%20one%20was%20right%20too%21%22%2C%22correct%22%3Atrue%7D%5D%2C%22randomize%22%3Afalse%2C%22multipleSelect%22%3Atrue%2C%22displayCount%22%3Anull%2C%22hasNoneOfTheAbove%22%3Afalse%2C%22deselectEnabled%22%3Afalse%7D%2C%22version%22%3A%7B%22major%22%3A1%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22calculator%22%3Afalse%2C%22chi2Table%22%3Afalse%2C%22periodicTable%22%3Afalse%2C%22tTable%22%3Afalse%2C%22zTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D)
 - Select a correct and an incorrect choice and click the "check answer" button
 - See that the correct choice stays selected and the incorrect choice becomes
   deselected.
 - Try selecting the previously-selected incorrect choice, see that I can't
 - Try de-selecting the still-selected correct choice, see that I can't
 - Select the other correct choice and click the "check answer" button
 - See that both of the correct choices stay selected
 - Try clicking all of the choices, see that they are now all read-only

Reviewers: audrey, kevinb

Reviewed By: audrey

Subscribers: lauren, jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D36636