Removing unused escape sequences from our JIPT hook.
Summary:
The characters that are represented by the escape sequences `\a`, `\b`,
`\f`, and `\v` are never used in our strings, yet our JIPT hook replaces
them. In combination with the "partial unescaping" JIPT bug and
Crowdin's partial change to "new-style" (not escaped) strings, this
caused the infamous "lue" bug in some strings, and made it impossible to
translate some new-style strings. Crowdin recently fixed the "partial
unescaping" bug (and introduced another one at the same time), but this
change should make it possible to translate most "new-style" strings.
Note that "new-style" strings that contain LaTeX macros that start with
`\n`, `\r`, or `\t` (the escape sequences we still have to replace
because we do use the corresponding characters in many strings) will
still be untranslatable.

This is a stop-gap measure while we await the Great Conversion Of All
Strings To The New Style.

Test Plan: - Browse around in the translation portal and see that "new-style" strings (the ones that only have single backslashes) are now translatable: the translation should contain single backslashes and should render properly without the "lue" bug. Which strings are "new-style" depend on when you downloaded your translation data. For example, on production, https://www.khanacademy.org/translations/edit/zh-hans/math/early-math/cc-early-math-place-value-topic/cc-early-math-two-digit-compare/e/comparing-two-digit-numbers-1/by-order currently contains a new-style string (type 3, problem 18 of 20, starting with "42 is 4 tens and 2 ones.")

Reviewers: emily, james, kevinb

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D37050