[cp mcr] Fix SAT review mode
Summary: With the new MCR changes, I changed some of the internal structure of
how we pass data into radio widgets. In particular, I always passed in the
correctness of each of the choices, where previously this was stripped out.
Unfortunately, this meant that when we restored serialized state on the widget
with data from before that change, the correcness is no longer passed in.

This fixes the problem by pulling that correctness from the review mode rubric,
like the widget used to do. This is a short-term fix though, and we should come
up with a more permanent fix for how to solve these serialized state issues in
the future.

Test Plan:
 - Visit http://localhost:8081/mission/sat/review
 - Open a review task that was done before the new MCR changes went out
 - Click "Review your answers"
 - See that one of the choices is marked as correct

Auditors: amy