[cp mcr] Handle choice focus state better
Summary: We use the `isInputFocused` state to determine when to show the focus
ring around some of our elements. We detect `blur` and `focus` events to turn
on and off this piece of state. Unfortunately, when an input becomes disabled,
it stops sending these events, which causes us to leave the state in whatever
it was before the input was disabled. When it is left in the "on" state, this
becomes confusing.

This change detects when we are about to set our input to disabled, and turns
off the focus state when we do. This prevents us from ever ending up in the
"focused but disabled" state.

Asana tasks:
https://app.asana.com/0/319778198354727/375216209278625
https://app.asana.com/0/319778198354727/375216209278623

Test Plan:
 - `make watch`
 - Visit
   http://localhost:8081/science/physics/one-dimensional-motion/acceleration-tutorial/e/rationales-test?final_mc_stylesk=1
 - Choose an incorrect choice
 - Press the enter button
 - See that the focus ring around the incorrect choice goes away

Auditors: amy, audrey