[cp mcr] Highlight all of the selected choices, not just new ones
Summary: I mis-interpreted the designs about which choices we wanted to be
highlighted during the mobile backdrop view. Apparently, we want all of the
choices that are selected to be highlighted, not just the new ones. This fixes
that by changing the state change and no longer keeping track of which ones
used to be highlighted.

Asana task:
https://app.asana.com/0/319778198354727/375298266423712

Test Plan:
 - Visit
   http://localhost:8081/science/physics/one-dimensional-motion/acceleration-tutorial/e/rationales-test?final_mc_stylesk=1
   in mobile emulation mode
 - Find a multi-select question
 - Choose one of the correct answers
 - Check your answer, see the selected choice is highlighted
 - Close the popover and also choose one of the incorrect answers
 - Check your answer, see that both of the selected choices are highlighted

Auditors: audrey