Deep compare widget trees, and only when the linter is on.
Summary:
This fixes a bug in a previous commit. I needed to force the linter to run
when the user edited a widget, but doing an === compare on the old and new
widget trees was apparently too permissive in some situations (such as SAT)
and we ended up re-rendering so frequently that MathJax couldn't keep up and
would never display its expressions. I don't understand why that was happening
but apparently the fix is to use _.isEqual() to deep compare the widget trees.
See https://khanacademy.slack.com/archives/C4WKMSC73/p1498973626652808

Also, my original commit did the comparison when the linter was off, which
was unnecessary, so now this change should never be able to affect student
facing exercises and articles. It only applys when we're in the editor with
the linter on.

Reviewers: emily, scottgrant, amy

Reviewed By: amy

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D37176