Hacky way to handle the fact that Crowdin escapes some, but not all, LaTeX strings.
We are currently in a situation where Crowdin adds extra backslashes
to some strings, but not all. Luckily, after a bugfix that is expected
for the week of 6/12/17, we will no longer see JIPT's partial unescaping
of some escaped strings. This will allow us to use a heuristic to determine
with a high probability whether a string is escaped or not, and thus
whether to unescape it or not in our `before_dom_insert()` JIPT hook.
The heuristic is described in `src/jipt-hack.jsx`.

The development and primary validation of the heuristic was done in Python
(the unit tests here are just to ensure that the same heuristic has been
correctly implemented in JavaScript). The Python code is included in this diff,
but will be removed before landing. In order to test the heuristic, `cd` into
the Perseus folder and download
`/home/ubuntu/i18n-update-strings/genfiles/translations/all.pot.txt_for_debugging`
from Jenkins into the Perseus folder and rename it to `all.pot`. Then, run
`PYTHONPATH=~/khan/webapp python jipthack/convert_pofile.py` in order to create
a file that contains the original strings in a simpler format and another file
that contains the result of escaping all of the original strings. Then, run
`python jipthack/heuristic.py` to see the outcomes of the heuristic.

This code is not expected to live for more than a few months (until we have
migrated all Crowdin strings to the "no escaping" form).

Test Plan:
- `./node_modules/.bin/mocha --reporter spec -r node/environment.js src/__tests__/jipt-hack-test.js`
- `make serve`
- Check that translations in still look good in http://localhost:8081/translations and http://localhost:8081/?lang=en-pt. In particular, find an item that has _not_ been escaped by Crowdin, and check that the translation renders correctly. For example, "42 is 4 tens and 2 ones" in problem 18 of type 3 in http://localhost:8081/translations/edit/zh-hans/math/early-math/cc-early-math-place-value-topic/cc-early-math-two-digit-compare/e/comparing-two-digit-numbers-1/by-order is not escaped, and has previously caused the "_lue_" bug.

Reviewers: emily, james

Reviewed By: emily, james

Differential Revision: https://phabricator.khanacademy.org/D36495