Removes eslint-ignore and runs prettier with some custom formatting
Summary:
This commit runs prettier over the rest of our Perseus components
directory and removes the bulk eslint-ignore warnings for style
issues.

`prettier --tab-width 4 --trailing-comma es5 --bracket-spacing false`

Again, it's pretty good, but not perfect! The single fix I've had to
address manually (except for console warnings and real lint issues
like props and React ordering) is the ternary formatting. Seems
like the prettier team is aware of this:

https://github.com/prettier/prettier/pull/1171

Otherwise, it's a lot of `var` to `let` or `const`. :)

Test Plan:
* `ka-lint`
* `make test`
* Test in a browser

Reviewers: emily

Reviewed By: emily

Subscribers: jared, kevinb, john

Differential Revision: https://phabricator.khanacademy.org/D37087