Adjusted definition box so that it appears above the word if the word is too low on the screen.
Summary: When the user first clicks or hovers over the word, we want the entire definition to be visible on the screen. Previously, the definition box always appeared underneath the word, regardless of where the word was on the screen, so sometimes the definition would be cut off at the bottom. So, I added a check that places the definition div above the word if the word is too low.

Test Plan:
 - Create a question and add 2 definition widgets: one that is near the top and one that is right above the bottom of the display box or screen.
 - (on desktop view) hover over the first word and verify that the definition box appears below and that all of the definition is visible.
 - still on desktop view, hover over the second word and verify that the definition box appears above the word and that all of the definition is visible.
 - Do the same for mobile view, except instead of hovering, you need to tap/click

You may need to add some more text at the bottom for the next steps.
 - Scroll so that the last defined word is at the bottom.
 - Click/hover over the word and verify that the definition box appears above.
 - Scroll so that the word is now in the middle of the screen. If on mobile, verify that the definition box stays open while scrolling, and close the definition box after you have scrolled the word to the middle of the screen.
 - Click/hover over the word to open the definition box again and verify that this time, the definition appears below.

Reviewers: emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D37441