[cp fixit] Hide the plotter "Drag handles to make graph" prompt after first interaction
Summary:
Before this change, we had the plotter decide to show the "Drag
handles to make graph" prompt if all of the values in the plotter were 0. This
was problematic for several reasons: on mobile, the default starting values are
not 0, and also content creators can change what the default starting values
are. This meant that in many cases users would probably never see the prompt.

This changes that behaviour by making the prompt show up at the beginning of
the problem regardless of values, and disappear after the first interaction
with the widget. This should ensure that people will always see the
instructions.

Asana task: https://app.asana.com/0/253524899192327/224784989615974

Screenshot:
{F772838}

Test Plan:
 - Visit [this rendered item with a plotter and non-0 starting values](http://localhost:9000/?renderer#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20plotter%201%5D%5D%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22plotter%201%22%3A%7B%22type%22%3A%22plotter%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22correct%22%3A%5B1%2C1%5D%2C%22starting%22%3A%5B6.5%2C4%5D%2C%22type%22%3A%22line%22%2C%22labels%22%3A%5B%22%22%2C%22%22%5D%2C%22categories%22%3A%5B%22a%22%2C%22b%22%5D%2C%22scaleY%22%3A1%2C%22maxY%22%3A10%2C%22snapsPerLine%22%3A2%2C%22labelInterval%22%3A1%7D%2C%22version%22%3A%7B%22major%22%3A0%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22calculator%22%3Afalse%2C%22chi2Table%22%3Afalse%2C%22periodicTable%22%3Afalse%2C%22tTable%22%3Afalse%2C%22zTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D)
   in mobile emulation mode
 - See that the "Drag handles to make graph" prompt is visible
 - Drag one of the handles, see that the prompt disappears.

Reviewers: charlie, scottgrant

Reviewed By: scottgrant

Subscribers: jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D37696