[cp fixit] Fix double-newline in code block jipt behaviour
Summary:
When we're in JIPT-mode, perseus tries to make sure that a translator
isn't doing anything bad to our translated paragraphs. We have to be careful to
continue rendering the same number of paragraph tags, otherwise we won't be
able to find the correct elements in the future. So, when content contained
double-newlines (which is generally the separator between paragraphs), we would
throw up a warning that someone was trying to split a paragraph in two.

In D35761, we added the ability for our perseus translation tools to understand
code blocks, because splitting a code block up into multiple paragraphs was
bad. Unfortunately, perseus didn't get the memo, and kept thinking that code
blocks with double-newlines were translators trying to split up paragraphs. -_-

This change makes perseus split content in the same way that the python does,
and also notice when there is a code block where it thinks a paragraph should
be and not give the warning.

Asana task: https://app.asana.com/0/253524899192327/325186023330295

Screenshots:
{F776205}
{F776208}

Test Plan:
 - `make watch`
 - Visit http://localhost:8081/translations/edit/lol/computing/computer-programming/programming-games-visualizations/programming-buttons/a/a-button-object-type/by-order?lang=en-pt
 - Log in to crowdin and use the LOLCAT language
 - See that the translated code blocks with newlines in them don't render as
   red "Please translate each paragraph to a single paragraph", but instead
   actually render as code blocks.
 - Translate one of the code blocks, see that it successfully updates, renders,
   and translates.
 - Add a double newline inside of a paragraph translation, see that it still
   gives me the "Please translate each paragraph to a single paragraph"
   error (as it should!)
 - Mess around with adding text before and after code blocks and putting
   incomplete code blocks to see if I can make the paragraph split into two.
   See that I can't! :)

Reviewers: johnsullivan, james

Reviewed By: james

Subscribers: jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D37738