[cp fixit] Fix JIPT rendering of passage widgets
Summary:
When doing JIPT (just in place translations), we need to have special
handling of the rendered text to make sure that the jipt code can find its jipt
content tags. We don't want to add support for all of that in the passage
widget and markdown, so when we see that we're in JIPT mode, we just pass off
that text to a normal `<Renderer />` which will handle things for us.

Screenshots:
{F776489}

Asana task:
https://app.asana.com/0/253524899192327/354966666506309

Test Plan:
 - `make watch`
 - Visit http://localhost:8081/translations/edit/lol/humanities/world-history/ancient-medieval/classical-greece/e/focus-on-state-building--the-greek-polis?return_to=/translations/lol/humanities/e&lang=en-pt
 - See that the passage widget shows up with the passage text
 - Click on it, see that the content can be translated and that it renders
   correctly.

Reviewers: james, mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D37752