If static, show the correct answer & rationale in the Radio (MC) widget
Summary:
From https://app.asana.com/0/253524899192327/385370526892703:
"The static setting of the MC widget isn't showing the correct answer selected.
This is problematic in history (and some other subject) exercises because the final
hint is supposed to show the final answer, but this bug is preventing the user from
seeing the correct answer."

Note that in this bug report, the widget in the hints is duplicated,
and does not have hints. (Should we better support this pattern? It seems
to be quite common in History...)

Test Plan:
 - In /gandalf, turn on content_library_rationales and final_multiple_choice_styles
 - Go to /humanities/ap-us-history/period-1/apush-old-and-new-worlds-collide/e/christopher-columbus-and-motivations-for-european-conquest
 - Go to the second hint, and see the answer displayed
 - Add a rationale to that exercise
 - In the perseus preview, load both hints and see the rationale displayed
 - {F768839}
 - Make sure that it works in the intermediate style as well.

Reviewers: emily

Reviewed By: emily

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D37686