Build Perseus.
Note that there were the following warnings:

```Warning: Unknown prop `keypadElement` on <input> tag. Remove this
prop from the element. For details, see https://fb.me/react-unknown-prop
    in input (created by Table)
    in td (created by Table)
    in tr (created by Table)
    in tbody (created by Table)
    in table (created by Table)
    in Table (created by WidgetContainer)
    in div (created by WidgetContainer)
    in WidgetContainer (created by Renderer)
    in div
    in div (created by QuestionParagraph)
    in QuestionParagraph (created by Renderer)
    in div (created by Renderer)
    in Renderer
Warning:
Unknown props `onFormatChange`, `checkValidity`, `useArrowKeys` on
<input> tag. Remove these props from the element. For details, see
https://fb.me/react-unknown-prop
    in input (created by NumberInput)
    in NumberInput
Warning: You are manually calling a React.PropTypes
validation function for the `value` prop on `<function tryPropType>`.
This is deprecated and will not work in the next major version. You may
be seeing this warning due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.
․Warning: You are manually calling a React.PropTypes validation function
for the `value` prop on `<function tryPropType>`. This is deprecated and
will not work in the next major version. You may be seeing this warning
due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.
․Warning: You are manually calling a React.PropTypes validation function
for the `value` prop on `<function tryPropType>`. This is deprecated and
will not work in the next major version. You may be seeing this warning
due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.
․Warning: You are manually calling a React.PropTypes validation function
for the `value` prop on `<function tryPropType>`. This is deprecated and
will not work in the next major version. You may be seeing this warning
due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.
․Warning: You are manually calling a React.PropTypes validation function
for the `value` prop on `<function tryPropType>`. This is deprecated and
will not work in the next major version. You may be seeing this warning
due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.
․Warning: You are manually calling a React.PropTypes validation function
for the `value` prop on `<function tryPropType>`. This is deprecated and
will not work in the next major version. You may be seeing this warning
due to a third-party PropTypes library. See
https://fb.me/react-warning-dont-call-proptypes for details.```