Adds context for content type and content domains for Gorgon
Summary:
We want to be able to write rules that are aware of the type
of content and of the domains that a content exists within.
This WIP adds two pieces of data to the Gorgon context object
so we can check if a single rule should be applied.

Test Plan:
* Test in browser
* More tests to follow, this is WIP at the moment

Reviewers: davidflanagan

Reviewed By: davidflanagan

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D38277