Flip the tooltip position based on its vertical position
Summary:
The lint tooltip isn't visible if the warning circle is too high
on the iframe. This commit finds the y-position of the circle,
and if it's too high on the page, uses a slightly different
positioning to place the tooltip below the circle instead. We
lose the cute tail (for now?), but it's usable.

Please don't make me find the CSS rules to make the tail show
up again. :) Consider that a feature request for the future?

I added a todo about the magic number 100. We can't use the
bounding rectangle function in render, according to React, so
maybe this is just something we can observe and tune as
needed. Maybe there's a better way too! For now, this will
let content creators see the tooltip if it's at the top.

Test Plan:
* Tested in browser
* `make test`

Reviewers: davidflanagan

Reviewed By: davidflanagan

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D38454