expression-editor: Fix answer deletion bug
Summary:
Deleting an answer option from the expression editor would always
delete the last answer, regardless of which answer you clicked
delete on.

This patch makes it so that clicking delete on an answer actually
deletes that answer option.

The primary bug is in handleRemoveForm, which would always remove
the last item. However, fixing just this meant that non-last
items can be deleted, which invalidates the expression-editor's
previous key generation scheme (of just using answerForms.length).

Therefore, we need to also write a function for generating a
correct key. So we look through all the answers and find a key
that is not already used, or return answerForms.length if none
are found.

Test plan:

1. added an expression widget
2. made answers 'a', 'b', 'c'
3. deleted answer 'b'.
4. saw that answers were now 'a' and 'c' (instead of 'a' and 'b'
   before this change)
5. reordered the answers
6. added a 'd'
7. checked the question's json to verify the keys were unique