Adds an applies function to rules for conditional checking
Summary:
Extends the rule class to allow for a function called
`applies` to check context for relevance. If we only want to
apply a rule in a particular content type or domain, we can
write a function to check for this.

A simple whitespace rule is added as an example that only
applies to articles. David, this rule should probably apply
on all types, but this is so you can see how I imagine this
working inside of new rules. Existing rules default to
applying in all content types, so there should be no
change in behaviour for existing content.

Note that for this rule, we can't check for newlines as
whitespace because they get stripped off when parsed into
a tree. We can still check for spaces at the end of content
though. I verified this in test and in browser.

Test Plan:
* Test in browser
* `make test`

Reviewers: davidflanagan

Reviewed By: davidflanagan

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D38450