Runs prettier and cleans up lint, just style changes
This hits a lot of code, but it's all style. So many style changes.

I'm tempted to just run prettier over the code and send an audit
instead. These thousand-line changes with no actual changes...

Summary:
...are
kind of a burden. :/ How about a one-and-done that can be
reverted if it causes a problem (unlikely, since it's just style)?

Test Plan:
* `ka-lint`
* `make test`
* Test in browser

Reviewers: joshuan, michaelpolyak

Reviewed By: joshuan

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D38749