Adds rule severity to Gorgon and visual circle indication
Summary:
This commit adds a new property to rules called `severity`.
It uses a number value defined in the `Rule` class to indicate
how important a rule is, and whether or not it should be shown
in the live editor or held for offline bulk-linter use.

There are some CSS styles in the `Lint` component that still
use the old orange colour. I'm not too concerned with that, and
certainly don't feel like it's breaking. Those rules are
embedded awfully deeply though, and it feels like pulling them
out will add a lot of complexity.

I also think the makeRule constructor is getting a little complex
too! Maybe we should be passing in an object instead of a
collection of arguments? I'd like to do that in a future commit
if it makes sense though, not here.

Test Plan:
* `ka-lint`
* `make test`
* Test in browser, verify colour changes, and that bulk linter
rules aren't displayed in the in-browser editor.

Reviewers: davidflanagan, emily

Reviewed By: davidflanagan

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D38750