[cp support] Fix correctness text in radio a11y text
Summary:
In SAT styling, we add some accessibility text to give users with
screenreaders more info about what is visually being shown on the screen. In
particular, we tell them whether each of the choices is correct or not.

In the past, we indicated whether or not a choice had been "checked" based on
whether the `correct` prop was `null` or not. However, in MCR I changed that so
`correct` was always non-null, but there was another prop, `showCorrectness`
which would say whether the choice had been "checked".

This diff stops depending on the nullable behavior of `correct` and uses
`showCorrectness` in those places.

Asana task: https://app.asana.com/0/265849773833386/415652218654789

Test Plan:
 - `make watch`
 - Visit http://localhost:8081/mission/sat/practice/math
 - Open a practice task
 - Inspect the choice icons, see that none of them say "correct" or "incorrect"
 - Answer a question correctly, inspect the choice icon and see that it now
   says "correct"
 - Go to the next question and answer it incorrectly. See that the selected
   choice says "incorrect" and the actual correct answer says "correct"

Reviewers: mattdunnrankin, amy

Reviewed By: mattdunnrankin

Subscribers: jared, #contentplatform

Differential Revision: https://phabricator.khanacademy.org/D38959