README: Add a link to perseus-configured
I made a repo to include all of perseus' global dependencies, and demo
how to require that. This seems to be a lot of the questions we get in
issues, so maybe linking to this in the readme can help people there.

This commit adds a link to https://github.com/ariabuckles/perseus-configured
in the README so that people can more easily find all the globals necessary to
use this, and so that we can maybe link it in issue answers.