Removes a test that causes a hang with Node 8
Summary:
I tried running `make build` with Perseus and node v8.4.0, which hangs. I checked into it a little further, and it's hanging on this test. Specifically, it hangs after triggering a `keypress` event inside of `mathQuillInput` after adding an open parenthesis to the form. You can do a quick check by leaving the test in place but changing the value of `specificWrong` from `"(x+2)(x-2)"` to something like `0`. This invalidates the test, but it doesn't hang anymore!

This diff is a WIP, or maybe a PSA. You might have already seen this, or might already know this is where the problem is! If so, sorry for the spam. :) It is kind of cool to build with Node 8.

I wasn't able to dive into MathQuill to figure out why this is happening. This might be a result of how it automagically inserts a close-parenthesis into the text box when you add an open-parenthesis. I don't know why that would hang in later Node versions. :/ You'd have the best idea about whether or not to ignore this for now since it's nice to have the test, or to consider adding a TODO and removing the test to get the advantages of upgrading.

Here's where the test was originally added: https://phabricator.khanacademy.org/D11057#inline-84748

Test Plan:
* Make sure you're using a newer version of node: `nvm use v8`
* `make test` should pass all tests successfully and should not hang.
* `make build` should produce new files in the `build` directory and should not hang.
* Opening a few exercises in your browser should show that Perseus still renders things as expected. For example, `http://localhost:8080/math/basic-geo/basic-geo-angle/measure-angles/e/measuring_angles`

Reviewers: emily

Reviewed By: emily

Subscribers: tom, jared

Differential Revision: https://phabricator.khanacademy.org/D39340