Adds a checkbox to the iframe editor to allow top navigation
Summary:
This was a fix put in on the eve of LearnStorm 2017 when we discovered that our TypeForm content wasn't granting badges. While this approach worked with LearnStorm 2016, browser changes led to security updates that squashed our badge grants. We had a big chat about this on Slack in the perseus channel:

https://khanacademy.slack.com/archives/C550Q5AMS/p1505176049000190

That led to this fix:

https://phabricator.khanacademy.org/D39308

Now we'd like to revisit it so that it's set on an individual iframe basis.

This first commit adds a checkbox, saves the status along with the iframe widget, and then blatantly disregards the proprerty to retain our existing behaviour. The reason for this is that we don't want to break the LearnStorm content in the brief period between deploying and publishing the articles with the saved data. Our plan is to:

* Deploy this change which enables the checkbox but still enables the `allow-top-navigation` property on all iframes.
* Edit the LearnStorm articles with iframes and save/publish with the property set.
* Deploy a second change that enables the conditional here that's commented out (right below the TODO, checking to see if `this.props.allowTopNavigation` is `true`)

Test Plan:
* View LearnStorm Activity #1 (probably available locally here: `http://localhost:8080/devadmin/content/articles/x2b9682344ec92b8f`, but if not, search for "Activity #1: The truth about your brain"). Confirm that the iframe widgets have a new checkbox at the bottom that says "Allow top navigation (LearnStorm)", and that the checkbox saves properly when you change and hit the "Save" button at the top right of the screen.
* Confirm that each iframe, regardless of the state of the checkbox, adds an `allow-top-navigation` property to the iframe on the right side (the rendered preview iframe). Screenshots will be attached to the diff.

Reviewers: emily, mattdunnrankin

Reviewed By: emily

Subscribers: tom, jared, nickbreen

Differential Revision: https://phabricator.khanacademy.org/D39646