Added id to inputs and for to labels
Summary:
As part of the test-prep a11y audit, a violation was raised at the highest
severity regarding our use of `input` elements nested in their `label` elements.
It was indicated that this is insufficient for screen reader usage and that the
labels must still reference the id of the input to which they relate.

This change generates an id for an input if it has none by using the `pos` value
and then adds `for` to the corresponding label (when not in edit mode).

Test Plan:
- Launch devserver and use a radio/choice component, then preview and review the
  choices in the developer tools and see that the label and inputs have the
  appropriate `for` and `id` values

Reviewers: nickbreen, jared, mattdunnrankin

Reviewed By: nickbreen, mattdunnrankin

Subscribers: #gtp

Differential Revision: https://phabricator.khanacademy.org/D39877