Add NOTIFIED file so that gerald will stop silently failing
Summary:
I've learned that the gerald action pretends to succeed if an error is thrown 🤦‍♂️😬 because it's async, and a promise rejection, and we're not handling them.
Anyway, gerald is currently dying silently because no NOTIFIED file exists. This is a cheap way to make it start working again, but we should also fix gerald.

Test plan:
See that gerald actually works on this PR!