Trigger eslint-reset and jest-reset on package.json changes (#261)
## Summary:

When the Changeset tool builds the "Version Packages" PR, it bumps the `package.json` files in this repo as-needed. We have Github actions that run ESLint and Flow for changes. In some cases, we want a full run and so we have the concept of a "reset" to trigger a full run. Any file listed in the reset list triggers the full run.

We had `yarn.lock` in the reset list for ESLint and Flow, but not `package.json`. Because `yarn.lock` isn't changed when we bump versions during a Changeset release, we were missing the full run we'd like to have. 

This PR adds `package.json` to the ESLint and Flow reset list.

Issue: "none"

## Test plan:

I'll land this and then check that the "Version Packages" PR that is already open does a full run of both ESLint and Flow. :fingers-crossed:

Author: jeremywiebe

Reviewers: kevinbarabash, jaredly

Required Reviewers:

Approved By: kevinbarabash

Checks: ✅ codecov/project, ✅ Lint, Flow, and Test (ubuntu-latest, 16.x), ✅ Extract i18n strings (ubuntu-latest, 16.x), ✅ Cypress Coverage (ubuntu-latest, 16.x), ✅ Check for .changeset file (ubuntu-latest, 16.x), ✅ Jest Coverage (ubuntu-latest, 16.x), ✅ gerald, ✅ Check builds for changes in size (ubuntu-latest, 16.x)

Pull Request URL: https://github.com/Khan/perseus/pull/261