Remove remaining references to getKaTeX
I noticed that, while the perseus editor seemed to be relying on
getKaTeX in order to display KaTeX rendering errors, it was also
importing KaTeX directly, making the call to getKaTeX redundant.  Both
webapp and mobile implement getKaTeX to simply return a reference to
katex, so this change to the perseus editor shouldn't affect behavior.

Issue: https://khanacademy.atlassian.net/browse/LC-459

Test plan:

Deploy a webapp ZND with this version of perseus-editor. Verify that the
"KaTeX errors" accordion in the exercise editor still displays any KaTeX
rendering errors.