Keypad: Move the FRAC button before the add/subtract/multiply/divide block
When the PERCENT button isn't present, the FRAC button
seems to be out of focus order when tabbing through the
Keypad. Also, even if it shows up in left-to-right order
when the PERCENT button is there, it still divides the
add/subtract/multiply/divide block, which is not intuitive
for screenreader users.

In this PR, I fixed this by moving the FRAC button just
before the PLUS button.

I also added a story for Keypad that includes all the
buttons excluding the navigation pad.

Issue: https://khanacademy.atlassian.net/browse/LC-1015

Test plan:

Storybook
- Start storybook with `yarn start`
- Go to `/?path=/story/math-input-full-keypad--everything-minus-navigation-pad`
- Tab through all the keys within every keypad page
- The fraction key should receieve focus before the plus key on
  every page.