Stop propogation in math input field click in order to prevent keypad closing
Currently, if a user has the keypad open and they click into the
input field, it automatically closes the keypad. This is not
the expected behavior - clicking outside the keypad should close it
EXCEPT when the clicked element is the math input.

In this PR:
- The math input field's `onClick` includes an `e.stopPropogation()`
  in order for the keypad to not read that click outside itself.
  This prevents the keypad from closing when the input is clicked.

Issue: https://khanacademy.atlassian.net/browse/LC-1410

Test plan:
todo