Update ArticleRenderer to split sections into paragraphs in JIPT context
This is a continuation to https://github.com/Khan/perseus/pull/996, where "disabling" of the lint
rule was removed from the markdown parser in order to correctly render lists in JIPT context.

However, that change broke Article preview rendering in the Translation Editor,
due to a mismatch in paragrap indexes. Causing the `jipt.js` integration to not
be able to attribute rendered strings to translations on Crowdin. This is fixed in: https://github.com/Khan/webapp/pull/19917

With these two changes in place, the Translation Editor is now rendering markdown
lists correctly without any other regression.

Unfortunately this is not case for articles containing markdown list rendered in
the WYSIWYG. These are regressed due to the same paragraph index mismatch, as unlike
the Translation Editor, the JIPT WYSIWYG renders articles using `ArticleRenderer`
(same as normal page rendering.)

To fix this, we apply the same change of splitting article sections into pragraphs
when rendering in JIPT context. This enables the `jipt.js` integration to attribute
the strings in the DOM to translations on Crowdin.

Issue: https://khanacademy.atlassian.net/browse/CP-7983

Test Plan:
 - Open any article in any KA locale site and verify rendering is unaffected
 - Open any article in WYSIWYG and verify there are no regressions to rendering
 - Open article with markdown list in WYSIWYG and verify list is rendered correctly and there are no other regressions
   - `/science/biology/classical-genetics/variations-on-mendelian-genetics/a/multiple-alleles-incomplete-dominance-and-codominance?lang=en-pt`