Use arrow functions in GraphSettings
I didn't update all the methods to arrow functions
in GraphSettings when updating it to be a React class,
which is now causes `this.setState()` to fail because
`this` is not bound.

The easiest way to make sure the functions are bound
to `this` is to make them arrow functions.

Issue: https://khanacademy.atlassian.net/browse/LC-1702

Test plan:
- Go to http://localhost:6006/?path=/docs/perseus-editor-editorpage--docs
- Click the "Add a widget" dropdown
- Choose interactive graph
- Open the "interactive-graph 1" section
- Open the web console
- Make a change to every single text field, dropdown, button, or other input.
- Check the console to make sure there are no errors trying to call
  `this.setState()`