Style movable points and segments on Mafs graphs
See this slack thread for designs and discussion:
https://khanacademy.slack.com/archives/C067UM1QAR4/p1709847469823399

Issue: LEMS-1672

Test plan:

```
yarn dev
open http://localhost:5173
```

Check the "segment" box in the dropdown at the top of the screen to
switch to the Mafs version of the segment graph.

You should be able to move the points and line segment with the mouse
and keyboard. The hover and focus states of the points should be clearly distinct.
The line segment should be highlighted blue when focused.