Implement adjusting colors for locked points
Now that the interactive graph editor can add locked
points, we also need to make it so that they can update
the color of those points.

In this PR, I added a dropdown that allows for changing
the color. I set the default color to blue. I also
made is so that the actual color value isn't saved
as part of the locked figure styles, just the name.
This then gets converted to the wonder blocks color
later in the mafs graph. I did it this way to avoid
having to convert back and forth between the color name
and value in the color UI.

Issue: https://khanacademy.atlassian.net/browse/LEMS-1751

Test plan:
todo