Ensure default segments are positioned correctly on Mafs graphs
Previously, there was a discrepancy between our `getDefaultSegments`
function and the equivalent code in the legacy InteractiveGraph.
`getDefaultSegments` was not "normalizing" the points correctly (see the
test added in this commit for an explanation of why we do this).

This commit ports the legacy behavior over.

Issue: https://khanacademy.atlassian.net/browse/LEMS-1841

Test plan:

See the linked issue for repro steps. The bug should not repro!