Refactor getGradableGraph
...and avoid returning coords if hasBeenInteractedWith is false.

The new structure allows us to avoid returning the coords property when
it is set on the initial graph. If coords is set, it makes the question
appear to have been answered even if the user never interacted with it.

Issue: https://khanacademy.atlassian.net/browse/LEMS-1776

Test plan:

`yarn test`

Co-authored-by: Mark Fitzgerald <markfitzgerald@khanacademy.org>