Increase the size of MovablePoint hitboxes to 48px
This is a holdover from the legacy interactive graph. 44px is WCAG AAA,
but there's no reason not to go up to 48px since that's the existing
behavior.

Issue: https://khanacademy.atlassian.net/browse/LEMS-1895

Test plan:

View a segment graph in the dev UI. Inspect one of the point hitboxes.
It should be 48 by 48 pixels.