Clarify the role of our graph context
We use a React context to convey contextual information about an
interactive graph down the component tree. I think when this context was
introduced, we envisioned it being the one-stop shop for all information
about a graph, but it turns out that doesn't work well. The different
graph types have different state shapes, so getting state from the
context in the individual graph components would force us to add a bunch
of redundant conditionals to narrow the types.

As a result of this limitation, the context was only used in practice
for information that doesn't vary between the graph types.

This PR renames and restructures the context to clarify its role.

Issue: none

Test plan:

`yarn test`