Don't extend graphed lines all the way to the graph bounds
Our SVG drawings get clipped off at the bounds of the Mafs SVG
container. This was causing the arrowheads at the end of lines to get
clipped in certain circumstances, e.g. when a line intersects the edge
of the graph at a small acute angle.

The solution implemented in this commit is to draw the arrowheads a few
pixels inside the graph bounds. Note that the legacy interactive graph
takes a similar approach to solving this problem.

Issue: LEMS-1893

Test plan:

- `yarn dev`
- View the Mafs linear graphs at `http://localhost:5173`. Drag the lines
  around. The arrowheads should remain visible.