Clarify the code for initializing linear graphs
Previously, the code for initializing movable lines looked like it would
use `graph.coords` if that was provided, and fall back to default values
if not. That's not what it did, because the `??` operator has higher
precedence than the ternary operator. If any coords were passed, we'd
treat the graph as a linear-system graph.

I suspect the reason we didn't notice this bug is that `coords` is never
provided for `linear` or `ray` graphs. Linear and ray graphs always have
one line, and linear-system graphs always have two lines.

Issue: none

Test plan:

- `yarn dev`
- Go to `http://localhost:5173#flipbook`
- View the examples obtained by running each of:
  ```
  ag -l '"type":"linear"' data/questions/ | xargs cat | grep -v 'grapher 1' | pbcopy
  ag -l '"type":"ray"' data/questions/ | xargs cat | pbcopy
  ag -l '"type":"linear-system"' data/questions/ | xargs cat | pbcopy
  ```
  (run these one at a time; each one overwrites your clipboard)
- Verify that none of the Mafs graphs differ from the legacy version.