Replace interactive graph snapshot tests with Storybook stories
There is a problem with using Jest snapshot tests to test Mafs: Mafs
uses autoincrementing IDs for some of its elements, so when new tests
are added to a suite, the IDs change and the snapshot tests fail.

To work around this, I've removed the problematic Jest snapshot tests
and replaced them with Storybook stories.

These stories will be run as visual snapshot tests on Chromatic,
providing similar regression protection with hopefully less test
flakiness.

Issue: LEMS-1975

Test plan:

Watch CI run and accept the baselines for the new stories on Chromatic.