Upgrade to @khanacademy/wonder-blocks-timing@5 everywhere
Previously, we were using version 4 in the dev UI, which caused
TypeScript errors on CI. TypeScript was finding the types for v5 when it
should have been using the types for v4.

Issue: LEMS-2002

Test plan:

View an interactive graph in the dev UI flipbook and use the "check
answer" button. It should pop up a banner that goes away after a couple
seconds.