Make the cursor for the circle edge point a left-right arrow
You can only drag this particular point to the left and right. The new
cursor makes that apparent.

Issue: LEMS-1979

Test plan:

Paste this data into the flipbook:

```
{"content":"**Graph the circle $(x+2)^2+(y+1)^2=49$.**\n\n[[☃ interactive-graph 1]]","images":{},"widgets":{"interactive-graph 1":{"alignment":"default","graded":true,"options":{"backgroundImage":{"url":null},"correct":{"center":[-2,-1],"radius":7,"type":"circle"},"graph":{"type":"circle"},"gridStep":[1,1],"labels":["x","y"],"markings":"graph","range":[[-10,10],[-10,10]],"rulerLabel":"","rulerTicks":10,"showProtractor":false,"showRuler":false,"snapStep":[1,1],"step":[1,1]},"static":false,"type":"interactive-graph","version":{"major":0,"minor":0}}}}
```

The cursor should change to a left-right arrow when you hover over the
movable point that changes the size of the circle.