[Interactive Graph Editor] Add the ability to reorder locked figure settings
Added some icon buttons so that content authors can reorder the locked figure
settings. This can be used to make sure that one figure can render on top
of another figure as desired. (The lower they are in the settings, the more
in the front they are on the graph.)

Issue: https://khanacademy.atlassian.net/browse/LEMS-1951

Test plan:
todo