[Storybook: Editor Page] Add a working preview for the editor page in storybook
The preview iframe doesn't work in storybook, because in links to a local URL
to webapp when it's in prod. It has never worked in storybook.

We need to be able to see what the initial state of the questions on the editor
page would look like without making any changes to the actual preview (yet).

I've added a preview here that shows the question and the hints in the editor
page. This works for all widgets. It's sticky, so it follows as you scroll
down a long editor. You can also close it if it's taking too much space.

This is a prerequisite to viewing an interactive graph with initial coordinates.

Issue: https://khanacademy.atlassian.net/browse/LEMS-2051

Test plan:
http://localhost:6006/?path=/story/perseuseditor-editorpage--demo