[Interactive Graph Editor] Fix the editor preview for segments, points, and polygons
There was an issue where some of the custom settings for segments, points, and polygon
graphs were not updating the right-side storybook preview for interactive graphs
in the EditorPage stories.

The issue is that the `onChange` functions in `interactive-graph-editor.tsx` were
updating the `correct` field, but they were not updating the `graph` field.

Updating the `graph` field as well in this PR.

Issue: none

Test plan:
todo