Add `mafs.point` flag to Renderer prop types
Perseus renders point graphs with fixed numbers of points using Mafs
when the `mafs.point` flag is on. Previously, however, the apiOptions
type (used in the Renderer props) did not allow this flag to be passed.
This commit makes it so the flag can be passed without a type error.