Make Perseus widgets open by default (#1643)
## Summary:
This PR sets the default state of Perseus widgets to be open. Additionally, it introduces a new optional prop to the EditorPage component, which is passed down to the WidgetEditor. This prop allows the visibility of all widgets on a page to be controlled with a single variable, enabling them to be opened or closed simultaneously.

This work is a precursor to planned work for the v2 exercise editor in Webapp in which a switch will be added to provide the ability to toggle widget visibility across the entire page.

Issue: [CP-8491](https://khanacademy.atlassian.net/browse/CP-8491) 

## Test plan:
Add a temporary switch to the editor-page which toggles the widget's visibility to ensure the widget opens/closes as expected. Test in Perseus storybook.

https://github.com/user-attachments/assets/b6b480ba-52a6-4bed-8648-98884acaf92a

Used the npm snapshot in webapp to verify this works with multiple widgets on the page.

https://github.com/user-attachments/assets/c08c88d5-f64e-4587-8314-de21517b1cc2


[CP-8491]: https://khanacademy.atlassian.net/browse/CP-8491?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Author: anniegallagher

Reviewers: jeremywiebe, anniegallagher

Required Reviewers:

Approved By: jeremywiebe

Checks: ✅ Upload Coverage (ubuntu-latest, 20.x), ✅ Publish npm snapshot (ubuntu-latest, 20.x), ✅ Jest Coverage (ubuntu-latest, 20.x), ✅ Check builds for changes in size (ubuntu-latest, 20.x), ✅ Lint, Typecheck, Format, and Test (ubuntu-latest, 20.x), ✅ Cypress (ubuntu-latest, 20.x), ✅ Publish Storybook to Chromatic (ubuntu-latest, 20.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ✅ gerald

Pull Request URL: https://github.com/Khan/perseus/pull/1643