Attempt bug fix for LabelImageEditor (#1830)
## Summary:
There's a race condition in the editing experience. I actually think the root of the issue has to do with code outside of Perseus, but CP is currently rewriting the editor so I'm just trying to patch for now and hope the real issue is resolved in the future.

Details are written out in the ticket, but the tl;dr is that there's some kind of race condition when changing the image and getting the image size. The outcome of the race condition was that `imageUrl` would become `""` when immediately calling `onChange` with an updated measurement. So to be doubly sure, I'm sending `imageUrl` with the measurements now.

**Risk:** I think there's a risk for a different type of race condition:
1. User changes URL
2. We start to measure the image size (async since we have to load the image)
3. User changes the URL again
4. We finish measuring the image size and bubble up the original `imageUrl` overwriting the new URL

I think this is less likely than the current race condition, but I thought I'd write it out in case there's a problem in the future.

Issue: [LEMS-2583](https://khanacademy.atlassian.net/browse/LEMS-2583)

## Test plan:
Since it's a race condition, this one was really hard to repro outside of Webapp. [Here are the instructions to repro](https://khanacademy.atlassian.net/browse/LEMS-2583?focusedCommentId=342225) in Webapp and ideally once this ships we won't be able to repro anymore.

[LEMS-2583]: https://khanacademy.atlassian.net/browse/LEMS-2583?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Author: handeyeco

Reviewers: Myranae

Required Reviewers:

Approved By: Myranae

Checks: ✅ Publish npm snapshot (ubuntu-latest, 20.x), ✅ Lint, Typecheck, Format, and Test (ubuntu-latest, 20.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ✅ Cypress (ubuntu-latest, 20.x), ✅ Publish Storybook to Chromatic (ubuntu-latest, 20.x), ✅ Check builds for changes in size (ubuntu-latest, 20.x), ✅ gerald

Pull Request URL: https://github.com/Khan/perseus/pull/1830