[Chromatic] Disable snapshots globally, only keep regression stories
Right now, Chromatic snapshots are only supposed to run on files
that end with `.regression.stories.tsx`, but they seem to be running
on other stories as well in the github checks.

Instead of using the .yml to specify this, I'm disabling chromatic
using the `disableSnapshot` chormatic parameter globablly, and
manually enabling it in the one regression stories file we have.

Issue: https://khanacademy.atlassian.net/browse/LEMS-2579

Test plan:
Check the chromatic info in the github checks below.