Migrate PerseusWidgetsMap to use a merged interface to represent all available widgets (#1936)
## Summary:

Perseus is designed to be able to support widgets registered into it at runtime. These widgets can be built in a separate npm package or even in the hosting application. As a result, the way that we've defined the `widgets` prop of `PerseusRenderer` is too rigid and poses long-term maintenance issues. 

As I was working on slitting out the bits of validation from scoring that we'll still be able to do on the client-side, I faced the same challenge (we will have a map of widget IDs for validation). 

Working with Kevin Barabash, we figured out a way to define an interface that represents all registered widgets and then use some TypeScript types to generate the `PerseusWidgetsMap`. This allows us to extend the "known set" of widgets from outside of Perseus. 

Although this is not an immediate concern, it will become more and more troublesome to build maps of widget information that are hard-coded to our current set of widgets. 

Issue: "none"

## Test plan:

`yarn typecheck`
`yarn test`

I'll import this npm snapshot into webapp to check that all types are fine there also.

EDIT: I installed the snapshot package in webapp and it type checks! ✅

Author: jeremywiebe

Reviewers: jeremywiebe, SonicScrewdriver, kevinb-khan, catandthemachines, handeyeco, benchristel

Required Reviewers:

Approved By: SonicScrewdriver, kevinb-khan, catandthemachines, handeyeco

Checks: ✅ Publish npm snapshot (ubuntu-latest, 20.x), ✅ Cypress (ubuntu-latest, 20.x), ✅ Check builds for changes in size (ubuntu-latest, 20.x), ✅ Lint, Typecheck, Format, and Test (ubuntu-latest, 20.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ✅ Publish Storybook to Chromatic (ubuntu-latest, 20.x), ✅ gerald

Pull Request URL: https://github.com/Khan/perseus/pull/1936