[Locked Figure Labels] Update placeholder to include TeX $s
Now that TeX needs $s around it to render as TeX,
update the placeholder in the interactive graph editor
label settings to reflect that.

While I'm here, I also updated the starting text on
the locked label settings to be "label" so that it's
easier to prevent empty labels. Charlie also expressed
that this makes it easier to tell the labels' tab orders.

Issue: none

Test plan:
- http://localhost:6006/?path=/story/perseuseditor-widgets-interactive-graph--mafs-with-locked-figure-labels-all-flags
- Go to locked label's settings
- Clear the text field
- Confirm that the placeholder says "ex. $x^2$ or $\frac{1}{2}$"