Mock Widget for cleaning up Tests (#2072)
## Summary:
This work was done as part of the Numeric Input project. 

This PR introduces a new Mock Widget to be used for any tests that doesn't require testing the capabilities of a specific widget. I have updated as many tests and/or testdatas as I could find, except for a few key situations:

- The test requires that we test a specific widget, such as extract-perseus-data. (Although I would like to recommend that we break up and move these tests into the individual widget tests as all we need to do is import a single function.) 
- The testdata is being used for storybook

I have also updated as many occurrences of any remaining uses of `input-number `in our tests to `numeric-input`, in order to help smooth the path forward for the eventual return to the Input Number Conversion project.   

I have opted not to include the widget in our extra-widgets / general registration for now as it would require that I also create an editor for it, but a caveat of this approach means that this widget is not accessible for Webapp tests. I would be happy to hear any thoughts regarding this approach, as I wasn't sure if I should be globally registering this widget. 

Issue: LEMS-2615

## Test plan:
- ensure all tests pass

Author: SonicScrewdriver

Reviewers: SonicScrewdriver, jeremywiebe, mark-fitzgerald

Required Reviewers:

Approved By: jeremywiebe, mark-fitzgerald

Checks: ✅ Publish npm snapshot (ubuntu-latest, 20.x), ✅ Check builds for changes in size (ubuntu-latest, 20.x), ✅ Lint, Typecheck, Format, and Test (ubuntu-latest, 20.x), ✅ Cypress (ubuntu-latest, 20.x), ✅ Check for .changeset entries for all changed files (ubuntu-latest, 20.x), ✅ Publish Storybook to Chromatic (ubuntu-latest, 20.x)

Pull Request URL: https://github.com/Khan/perseus/pull/2072